Add progress bar to frames

This function adds a progress bar to frames created with frames_spatial.

Usage

add_progress(frames, colour = "grey", size = 1.8, y_nudge = 0, verbose = TRUE)

Arguments

frames

an object of class moveVis created with frames_spatial.

colour

character, progress bar colour.

size

numeric, progress bar line size (line width).

y_nudge

numeric, amount of vertical distance to move the progress bar (default: top of plot area).

verbose

logical, if TRUE, messages and progress information are displayed on the console (default).

Value

A frames object of class moveVis.

See also

Author

Jakob Schwalb-Willmann

Examples

library(moveVis)
library(move)
#> Error in library(move): there is no package called ‘move’

data("move_data", "basemap_data")
#> Warning: data set ‘basemap_data’ not found
m <- align_move(move_data, res = 4, unit = "mins")
#> Warning: Argument 'unit' is deprecated. See ?moveVis::align_move for details.
#> Error: Argument 'res' must either be a 'units' object or one of c('min', 'max', 'mean', 'median').

# create spatial frames using a custom NDVI base layer
r_list <- basemap_data[[1]]
#> Error: object 'basemap_data' not found
r_times <- basemap_data[[2]]
#> Error: object 'basemap_data' not found

if (FALSE) { # \dontrun{
frames <- frames_spatial(m, r_list = r_list, r_times = r_times, r_type = "gradient",
                         fade_raster = TRUE)

# add a progress bar:
frames.a <- add_progress(frames)
frames.a[[100]]

# or in red and larger
frames.b <- add_progress(frames, colour = "red", size = 2.5)
frames.b[[100]]
} # }