Add scalebar to frames

This function adds a scalebar to frames created with frames_spatial.

Usage

add_scalebar(
  frames,
  distance = NULL,
  height = 0.015,
  position = "bottomleft",
  x = NULL,
  y = NULL,
  colour = "black",
  label_margin = 1.2,
  units = "km",
  verbose = TRUE
)

Arguments

frames

an object of class moveVis created with frames_spatial.

distance

numeric, optional. Distance displayed by the scalebar (in either km or miles defined by argument units) By default, the displayed distance is calculated automatically.

height

numeric, height of the scalebar in a range from 0 to 1 as the proportion of the overall height of the frame map. Default is 0.015.

position

character, position of the scalebar on the map. Either "bottomleft", "upperleft", "upperright", "bottomright". Ignored, if x and y are set.

x

numeric, position of the bottom left corner of the scalebar on the x axis. If not set, position is used to calculate the position of the scalebar.

y

numeric, position of the bottom left corner of the scalebar on the y axis. If not set, position is used to calculate the position of the scalebar.

colour

character, colour of the distance labels. Default is "black".

label_margin

numeric, distance of the labels to the scalebar as a proportion of the height of the scalebar (e.g. if set to 2, the labels will be positioned with a distance to the scalebar of twice the scalebar height).

units

character, either "km" for kilometers or "miles" for miles.

verbose

logical, if TRUE, messages and progress information are displayed on the console (default).

Value

A frames object of class moveVis.

Details

For adding more sophisticated map features, I recommend using ggspatial.

See also

Author

Jakob Schwalb-Willmann

Examples

library(moveVis)
library(move)
#> Error in library(move): there is no package called ‘move’

data("move_data", "basemap_data")
#> Warning: data set ‘basemap_data’ not found
m <- align_move(move_data, res = 4, unit = "mins")
#> Warning: Argument 'unit' is deprecated. See ?moveVis::align_move for details.
#> Error: Argument 'res' must either be a 'units' object or one of c('min', 'max', 'mean', 'median').

# create spatial frames using a custom NDVI base layer
r_list <- basemap_data[[1]]
#> Error: object 'basemap_data' not found
r_times <- basemap_data[[2]]
#> Error: object 'basemap_data' not found

if (FALSE) { # \dontrun{
frames <- frames_spatial(m, r_list = r_list, r_times = r_times, r_type = "gradient",
                         fade_raster = TRUE)

# add a scale bar to frames:
frames.a <- add_scalebar(frames)
frames.a[[100]]

# or in white at another position
frames.b <- add_scalebar(frames, colour = "white", position = "bottomright")
frames.b[[100]]

# or with another height
frames.c <- add_scalebar(frames, colour = "white", position = "bottomright", height = 0.025)
frames.c[[100]]
} # }