Align movement data

This function aligns movement data to a uniform time scale with a uniform temporal resolution throughout the complete movement sequence. This prepares the provided movement data to be usable by frames_spatial, which requires a uniform time scale and a consistent, unique temporal resolution for all supplied trajectories.

Usage

align_move(
  m,
  res = "minimum",
  start_end_time = NULL,
  fill_na_values = TRUE,
  ...,
  verbose = TRUE
)

Arguments

m

move2 object, which is allowed to contain irregular timestamps and diverging temporal resolutions.

res

either a units object representing the temporal resolution m should be aligned to, or a character being one of 'minimum', 'maximum', 'mean' or 'median' to indicate how the target resolution should be derived from m.

start_end_time

NULL (default) or a vector of two POSIXct times (one start time and one end time for alignment). If NULL, the start and end time are retrieved from m and used for alignment.

"minimum" to use the smallest temporal resolution of m (default)

"maximum" to use the largest temporal resolution of m

"mean" to use the rounded average temporal resolution of m

"median" to use the rounded median temporal resolution of m

fill_na_values

logical, whether to fill empty (NA) values of columns of m at interpolated locations (defaults to TRUE). Column values at interpolated locations are filled with the value of the temporally closest location.

...

deprecated arguments, including digit, unit and spaceMethod.

verbose

logical, if TRUE, messages and progress information are displayed on the console (default).

Value

move2 object, with aligned positions at uniform temporal scale computed from m, ready to be used by frames_spatial.

See also

Author

Jakob Schwalb-Willmann

Examples

library(moveVis)
library(move2)
library(lubridate)
#> 
#> Attaching package: ‘lubridate’
#> The following objects are masked from ‘package:terra’:
#> 
#>     intersect, union
#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

# example data
data("move_data")

# the tracks in move_data have irregular timestamps and sampling rates.
# print unique timestamps and time lags
unique(mt_time(move_data))
#>   [1] "2018-05-15 09:59:59 UTC" "2018-05-15 10:04:01 UTC"
#>   [3] "2018-05-15 10:08:02 UTC" "2018-05-15 10:12:04 UTC"
#>   [5] "2018-05-15 10:16:05 UTC" "2018-05-15 10:20:06 UTC"
#>   [7] "2018-05-15 10:24:07 UTC" "2018-05-15 10:28:07 UTC"
#>   [9] "2018-05-15 10:32:10 UTC" "2018-05-15 10:36:12 UTC"
#>  [11] "2018-05-15 10:40:11 UTC" "2018-05-15 10:44:14 UTC"
#>  [13] "2018-05-15 10:48:13 UTC" "2018-05-15 10:52:16 UTC"
#>  [15] "2018-05-15 10:56:17 UTC" "2018-05-15 11:00:18 UTC"
#>  [17] "2018-05-15 11:04:20 UTC" "2018-05-15 11:08:20 UTC"
#>  [19] "2018-05-15 11:12:21 UTC" "2018-05-15 11:16:22 UTC"
#>  [21] "2018-05-15 11:20:24 UTC" "2018-05-15 11:24:26 UTC"
#>  [23] "2018-05-15 11:28:25 UTC" "2018-05-15 11:32:29 UTC"
#>  [25] "2018-05-15 11:36:30 UTC" "2018-05-15 11:40:31 UTC"
#>  [27] "2018-05-15 11:44:31 UTC" "2018-05-15 11:48:34 UTC"
#>  [29] "2018-05-15 11:52:35 UTC" "2018-05-15 11:56:39 UTC"
#>  [31] "2018-05-15 12:00:38 UTC" "2018-05-15 12:04:38 UTC"
#>  [33] "2018-05-15 12:08:41 UTC" "2018-05-15 12:12:43 UTC"
#>  [35] "2018-05-15 12:16:43 UTC" "2018-05-15 12:20:44 UTC"
#>  [37] "2018-05-15 12:24:45 UTC" "2018-05-15 12:28:44 UTC"
#>  [39] "2018-05-15 12:32:48 UTC" "2018-05-15 12:36:48 UTC"
#>  [41] "2018-05-15 12:40:50 UTC" "2018-05-15 12:44:51 UTC"
#>  [43] "2018-05-15 12:48:53 UTC" "2018-05-15 12:52:53 UTC"
#>  [45] "2018-05-15 12:56:55 UTC" "2018-05-15 13:00:56 UTC"
#>  [47] "2018-05-15 13:04:58 UTC" "2018-05-15 13:08:59 UTC"
#>  [49] "2018-05-15 13:13:00 UTC" "2018-05-15 13:17:01 UTC"
#>  [51] "2018-05-15 13:21:03 UTC" "2018-05-15 13:25:02 UTC"
#>  [53] "2018-05-15 13:29:03 UTC" "2018-05-15 13:33:05 UTC"
#>  [55] "2018-05-15 13:37:07 UTC" "2018-05-15 13:41:07 UTC"
#>  [57] "2018-05-15 13:45:11 UTC" "2018-05-15 13:49:12 UTC"
#>  [59] "2018-05-15 13:53:13 UTC" "2018-05-15 13:57:15 UTC"
#>  [61] "2018-05-15 14:01:15 UTC" "2018-05-15 14:05:17 UTC"
#>  [63] "2018-05-15 14:09:18 UTC" "2018-05-15 14:13:19 UTC"
#>  [65] "2018-05-15 14:17:19 UTC" "2018-05-15 14:21:21 UTC"
#>  [67] "2018-05-15 14:25:22 UTC" "2018-05-15 14:29:25 UTC"
#>  [69] "2018-05-15 14:33:26 UTC" "2018-05-15 14:37:25 UTC"
#>  [71] "2018-05-15 14:41:27 UTC" "2018-05-15 14:45:29 UTC"
#>  [73] "2018-05-15 14:49:30 UTC" "2018-05-15 14:53:31 UTC"
#>  [75] "2018-05-15 14:57:32 UTC" "2018-05-15 15:01:33 UTC"
#>  [77] "2018-05-15 15:05:35 UTC" "2018-05-15 15:09:38 UTC"
#>  [79] "2018-05-15 15:13:36 UTC" "2018-05-15 15:17:37 UTC"
#>  [81] "2018-05-15 15:21:39 UTC" "2018-05-15 15:25:42 UTC"
#>  [83] "2018-05-15 15:29:43 UTC" "2018-05-15 15:33:46 UTC"
#>  [85] "2018-05-15 15:37:46 UTC" "2018-05-15 15:41:47 UTC"
#>  [87] "2018-05-15 15:45:48 UTC" "2018-05-15 15:49:49 UTC"
#>  [89] "2018-05-15 15:53:52 UTC" "2018-05-15 15:57:51 UTC"
#>  [91] "2018-05-15 16:01:54 UTC" "2018-05-15 16:05:52 UTC"
#>  [93] "2018-05-15 16:09:55 UTC" "2018-05-15 16:13:56 UTC"
#>  [95] "2018-05-15 16:17:57 UTC" "2018-05-15 16:21:58 UTC"
#>  [97] "2018-05-15 16:26:02 UTC" "2018-05-15 16:30:02 UTC"
#>  [99] "2018-05-15 16:34:03 UTC" "2018-05-15 16:38:03 UTC"
#> [101] "2018-05-15 16:42:06 UTC" "2018-05-15 16:46:06 UTC"
#> [103] "2018-05-15 16:50:08 UTC" "2018-05-15 16:54:11 UTC"
#> [105] "2018-05-15 16:58:10 UTC" "2018-05-15 17:02:11 UTC"
#> [107] "2018-05-15 17:06:13 UTC" "2018-05-15 17:10:13 UTC"
#> [109] "2018-05-15 17:14:17 UTC" "2018-05-15 17:18:17 UTC"
#> [111] "2018-05-15 17:22:18 UTC" "2018-05-15 17:26:18 UTC"
#> [113] "2018-05-15 17:30:20 UTC" "2018-05-15 17:34:22 UTC"
#> [115] "2018-05-15 17:38:24 UTC" "2018-05-15 17:42:25 UTC"
#> [117] "2018-05-15 17:46:27 UTC" "2018-05-15 17:50:27 UTC"
#> [119] "2018-05-15 17:54:27 UTC" "2018-05-15 17:58:30 UTC"
#> [121] "2018-05-15 18:02:30 UTC" "2018-05-15 18:06:32 UTC"
#> [123] "2018-05-15 18:10:34 UTC" "2018-05-15 18:14:36 UTC"
#> [125] "2018-05-15 18:18:36 UTC" "2018-05-15 18:22:37 UTC"
#> [127] "2018-05-15 18:26:36 UTC" "2018-05-15 18:30:40 UTC"
#> [129] "2018-05-15 18:34:41 UTC" "2018-05-15 18:38:42 UTC"
#> [131] "2018-05-15 18:42:44 UTC" "2018-05-15 18:46:46 UTC"
#> [133] "2018-05-15 18:50:45 UTC" "2018-05-15 18:54:47 UTC"
#> [135] "2018-05-15 18:58:48 UTC" "2018-05-15 19:02:49 UTC"
#> [137] "2018-05-15 19:06:51 UTC" "2018-05-15 19:10:52 UTC"
#> [139] "2018-05-15 19:14:53 UTC" "2018-05-15 19:18:55 UTC"
#> [141] "2018-05-15 19:22:55 UTC" "2018-05-15 19:26:57 UTC"
#> [143] "2018-05-15 19:30:57 UTC" "2018-05-15 19:34:59 UTC"
#> [145] "2018-05-15 08:00:00 UTC" "2018-05-15 08:04:00 UTC"
#> [147] "2018-05-15 08:08:02 UTC" "2018-05-15 08:12:01 UTC"
#> [149] "2018-05-15 08:16:03 UTC" "2018-05-15 08:20:07 UTC"
#> [151] "2018-05-15 08:24:11 UTC" "2018-05-15 08:28:10 UTC"
#> [153] "2018-05-15 08:32:11 UTC" "2018-05-15 08:36:12 UTC"
#> [155] "2018-05-15 08:40:14 UTC" "2018-05-15 08:44:14 UTC"
#> [157] "2018-05-15 08:48:19 UTC" "2018-05-15 08:52:19 UTC"
#> [159] "2018-05-15 08:56:21 UTC" "2018-05-15 09:00:23 UTC"
#> [161] "2018-05-15 09:04:26 UTC" "2018-05-15 09:08:25 UTC"
#> [163] "2018-05-15 09:12:28 UTC" "2018-05-15 09:16:28 UTC"
#> [165] "2018-05-15 09:20:31 UTC" "2018-05-15 09:24:31 UTC"
#> [167] "2018-05-15 09:28:32 UTC" "2018-05-15 09:32:33 UTC"
#> [169] "2018-05-15 09:36:34 UTC" "2018-05-15 09:40:37 UTC"
#> [171] "2018-05-15 09:44:42 UTC" "2018-05-15 09:48:41 UTC"
#> [173] "2018-05-15 09:52:41 UTC" "2018-05-15 09:56:44 UTC"
#> [175] "2018-05-15 10:00:45 UTC" "2018-05-15 10:04:46 UTC"
#> [177] "2018-05-15 10:08:48 UTC" "2018-05-15 10:12:49 UTC"
#> [179] "2018-05-15 10:16:53 UTC" "2018-05-15 10:20:53 UTC"
#> [181] "2018-05-15 10:24:53 UTC" "2018-05-15 10:28:56 UTC"
#> [183] "2018-05-15 10:32:57 UTC" "2018-05-15 10:36:58 UTC"
#> [185] "2018-05-15 10:41:00 UTC" "2018-05-15 10:45:03 UTC"
#> [187] "2018-05-15 10:49:02 UTC" "2018-05-15 10:53:05 UTC"
#> [189] "2018-05-15 10:57:07 UTC" "2018-05-15 11:01:08 UTC"
#> [191] "2018-05-15 11:05:07 UTC" "2018-05-15 11:09:11 UTC"
#> [193] "2018-05-15 11:13:12 UTC" "2018-05-15 11:17:15 UTC"
#> [195] "2018-05-15 11:21:16 UTC" "2018-05-15 11:25:15 UTC"
#> [197] "2018-05-15 11:29:16 UTC" "2018-05-15 11:33:20 UTC"
#> [199] "2018-05-15 11:37:22 UTC" "2018-05-15 11:41:22 UTC"
#> [201] "2018-05-15 11:45:25 UTC" "2018-05-15 11:49:24 UTC"
#> [203] "2018-05-15 11:53:27 UTC" "2018-05-15 11:57:28 UTC"
#> [205] "2018-05-15 12:01:29 UTC" "2018-05-15 12:05:30 UTC"
#> [207] "2018-05-15 12:09:34 UTC" "2018-05-15 12:13:33 UTC"
#> [209] "2018-05-15 12:17:36 UTC" "2018-05-15 12:21:36 UTC"
#> [211] "2018-05-15 12:25:40 UTC" "2018-05-15 12:29:40 UTC"
#> [213] "2018-05-15 12:33:42 UTC" "2018-05-15 12:37:44 UTC"
#> [215] "2018-05-15 12:41:46 UTC" "2018-05-15 12:45:48 UTC"
#> [217] "2018-05-15 12:49:48 UTC" "2018-05-15 12:53:48 UTC"
#> [219] "2018-05-15 12:57:51 UTC" "2018-05-15 13:01:52 UTC"
#> [221] "2018-05-15 13:05:53 UTC" "2018-05-15 13:09:55 UTC"
#> [223] "2018-05-15 13:13:57 UTC" "2018-05-15 13:18:00 UTC"
#> [225] "2018-05-15 13:22:00 UTC" "2018-05-15 13:26:03 UTC"
#> [227] "2018-05-15 13:30:04 UTC" "2018-05-15 13:34:07 UTC"
#> [229] "2018-05-15 13:38:04 UTC" "2018-05-15 13:42:08 UTC"
#> [231] "2018-05-15 13:46:10 UTC" "2018-05-15 13:50:11 UTC"
#> [233] "2018-05-15 13:54:11 UTC" "2018-05-15 13:58:15 UTC"
#> [235] "2018-05-15 14:02:15 UTC" "2018-05-15 14:06:16 UTC"
#> [237] "2018-05-15 14:10:17 UTC" "2018-05-15 14:14:21 UTC"
#> [239] "2018-05-15 14:18:19 UTC" "2018-05-15 14:22:24 UTC"
#> [241] "2018-05-15 14:26:23 UTC" "2018-05-15 14:30:27 UTC"
#> [243] "2018-05-15 14:34:28 UTC" "2018-05-15 14:38:29 UTC"
#> [245] "2018-05-15 14:42:29 UTC" "2018-05-15 14:46:30 UTC"
#> [247] "2018-05-15 14:50:34 UTC" "2018-05-15 14:54:35 UTC"
#> [249] "2018-05-15 14:58:35 UTC" "2018-05-15 15:02:38 UTC"
#> [251] "2018-05-15 15:06:38 UTC" "2018-05-15 15:10:42 UTC"
#> [253] "2018-05-15 15:14:41 UTC" "2018-05-15 15:18:44 UTC"
#> [255] "2018-05-15 15:22:44 UTC" "2018-05-15 15:26:48 UTC"
#> [257] "2018-05-15 15:30:49 UTC" "2018-05-15 15:34:51 UTC"
#> [259] "2018-05-15 15:38:52 UTC" "2018-05-15 15:42:54 UTC"
#> [261] "2018-05-15 15:46:54 UTC" "2018-05-15 15:50:55 UTC"
#> [263] "2018-05-15 15:54:59 UTC" "2018-05-15 15:59:00 UTC"
#> [265] "2018-05-15 16:03:00 UTC" "2018-05-15 16:07:01 UTC"
#> [267] "2018-05-15 16:11:04 UTC" "2018-05-15 16:15:04 UTC"
#> [269] "2018-05-15 16:19:06 UTC" "2018-05-15 16:23:09 UTC"
#> [271] "2018-05-15 16:27:11 UTC" "2018-05-15 16:31:11 UTC"
#> [273] "2018-05-15 16:35:12 UTC" "2018-05-15 16:39:14 UTC"
#> [275] "2018-05-15 16:43:17 UTC" "2018-05-15 16:47:16 UTC"
#> [277] "2018-05-15 16:51:20 UTC" "2018-05-15 16:55:20 UTC"
#> [279] "2018-05-15 16:59:23 UTC" "2018-05-15 17:03:24 UTC"
#> [281] "2018-05-15 17:07:26 UTC" "2018-05-15 17:11:26 UTC"
#> [283] "2018-05-15 17:15:29 UTC" "2018-05-15 17:19:27 UTC"
#> [285] "2018-05-15 17:23:32 UTC" "2018-05-15 17:27:33 UTC"
#> [287] "2018-05-15 17:31:34 UTC" "2018-05-15 17:35:35 UTC"
#> [289] "2018-05-15 17:39:36 UTC" "2018-05-15 17:43:37 UTC"
#> [291] "2018-05-15 17:47:39 UTC" "2018-05-15 17:51:42 UTC"
#> [293] "2018-05-15 17:55:42 UTC" "2018-05-15 17:59:46 UTC"
#> [295] "2018-05-15 18:03:48 UTC" "2018-05-15 18:07:46 UTC"
#> [297] "2018-05-15 18:11:48 UTC" "2018-05-15 18:15:50 UTC"
#> [299] "2018-05-15 18:19:52 UTC" "2018-05-15 18:23:54 UTC"
#> [301] "2018-05-15 18:27:55 UTC" "2018-05-15 18:31:57 UTC"
#> [303] "2018-05-15 18:35:58 UTC" "2018-05-15 18:40:00 UTC"
#> [305] "2018-05-15 06:59:59 UTC" "2018-05-15 07:04:00 UTC"
#> [307] "2018-05-15 07:08:04 UTC" "2018-05-15 07:12:02 UTC"
#> [309] "2018-05-15 07:16:05 UTC" "2018-05-15 07:20:05 UTC"
#> [311] "2018-05-15 07:24:06 UTC" "2018-05-15 07:28:06 UTC"
#> [313] "2018-05-15 07:32:09 UTC" "2018-05-15 07:36:10 UTC"
#> [315] "2018-05-15 07:40:10 UTC" "2018-05-15 07:44:13 UTC"
#> [317] "2018-05-15 07:48:14 UTC" "2018-05-15 07:52:17 UTC"
#> [319] "2018-05-15 07:56:17 UTC" "2018-05-15 08:00:18 UTC"
#> [321] "2018-05-15 08:04:19 UTC" "2018-05-15 08:08:18 UTC"
#> [323] "2018-05-15 08:12:20 UTC" "2018-05-15 08:16:20 UTC"
#> [325] "2018-05-15 08:20:25 UTC" "2018-05-15 08:24:24 UTC"
#> [327] "2018-05-15 08:28:23 UTC" "2018-05-15 08:32:26 UTC"
#> [329] "2018-05-15 08:36:28 UTC" "2018-05-15 08:40:30 UTC"
#> [331] "2018-05-15 08:44:31 UTC" "2018-05-15 08:48:32 UTC"
#> [333] "2018-05-15 08:52:33 UTC" "2018-05-15 08:56:34 UTC"
#> [335] "2018-05-15 09:00:35 UTC" "2018-05-15 09:04:36 UTC"
#> [337] "2018-05-15 09:08:37 UTC" "2018-05-15 09:12:40 UTC"
#> [339] "2018-05-15 09:16:40 UTC" "2018-05-15 09:20:40 UTC"
#> [341] "2018-05-15 09:24:43 UTC" "2018-05-15 09:28:43 UTC"
#> [343] "2018-05-15 09:32:43 UTC" "2018-05-15 09:36:45 UTC"
#> [345] "2018-05-15 09:40:47 UTC" "2018-05-15 09:44:47 UTC"
#> [347] "2018-05-15 09:48:49 UTC" "2018-05-15 09:52:50 UTC"
#> [349] "2018-05-15 09:56:50 UTC" "2018-05-15 10:00:53 UTC"
#> [351] "2018-05-15 10:04:55 UTC" "2018-05-15 10:08:54 UTC"
#> [353] "2018-05-15 10:12:56 UTC" "2018-05-15 10:16:58 UTC"
#> [355] "2018-05-15 10:20:57 UTC" "2018-05-15 10:25:01 UTC"
#> [357] "2018-05-15 10:29:00 UTC" "2018-05-15 10:33:02 UTC"
#> [359] "2018-05-15 10:37:04 UTC" "2018-05-15 10:41:04 UTC"
#> [361] "2018-05-15 10:45:06 UTC" "2018-05-15 10:49:06 UTC"
#> [363] "2018-05-15 10:53:08 UTC" "2018-05-15 10:57:08 UTC"
#> [365] "2018-05-15 11:01:10 UTC" "2018-05-15 11:05:11 UTC"
#> [367] "2018-05-15 11:09:13 UTC" "2018-05-15 11:13:13 UTC"
#> [369] "2018-05-15 11:17:14 UTC" "2018-05-15 11:21:16 UTC"
#> [371] "2018-05-15 11:25:20 UTC" "2018-05-15 11:29:17 UTC"
#> [373] "2018-05-15 11:33:21 UTC" "2018-05-15 11:37:20 UTC"
#> [375] "2018-05-15 11:41:22 UTC" "2018-05-15 11:45:20 UTC"
#> [377] "2018-05-15 11:49:25 UTC" "2018-05-15 11:53:24 UTC"
#> [379] "2018-05-15 11:57:26 UTC" "2018-05-15 12:01:27 UTC"
#> [381] "2018-05-15 12:05:30 UTC" "2018-05-15 12:09:30 UTC"
#> [383] "2018-05-15 12:13:30 UTC" "2018-05-15 12:17:31 UTC"
#> [385] "2018-05-15 12:21:34 UTC" "2018-05-15 12:25:36 UTC"
#> [387] "2018-05-15 12:29:36 UTC" "2018-05-15 12:33:37 UTC"
#> [389] "2018-05-15 12:37:38 UTC" "2018-05-15 12:41:40 UTC"
#> [391] "2018-05-15 12:45:42 UTC" "2018-05-15 12:49:42 UTC"
#> [393] "2018-05-15 12:53:44 UTC" "2018-05-15 12:57:44 UTC"
#> [395] "2018-05-15 13:01:46 UTC" "2018-05-15 13:05:45 UTC"
#> [397] "2018-05-15 13:09:49 UTC" "2018-05-15 13:13:48 UTC"
#> [399] "2018-05-15 13:17:50 UTC" "2018-05-15 13:21:50 UTC"
#> [401] "2018-05-15 13:25:51 UTC" "2018-05-15 13:29:54 UTC"
#> [403] "2018-05-15 13:33:55 UTC" "2018-05-15 13:37:55 UTC"
#> [405] "2018-05-15 13:41:56 UTC" "2018-05-15 13:45:58 UTC"
#> [407] "2018-05-15 13:50:00 UTC" "2018-05-15 13:54:01 UTC"
#> [409] "2018-05-15 13:58:02 UTC" "2018-05-15 14:02:05 UTC"
#> [411] "2018-05-15 14:06:04 UTC" "2018-05-15 14:10:06 UTC"
#> [413] "2018-05-15 14:14:08 UTC" "2018-05-15 14:18:07 UTC"
#> [415] "2018-05-15 14:22:09 UTC" "2018-05-15 14:26:08 UTC"
#> [417] "2018-05-15 14:30:11 UTC" "2018-05-15 14:34:13 UTC"
#> [419] "2018-05-15 14:38:13 UTC" "2018-05-15 14:42:13 UTC"
#> [421] "2018-05-15 14:46:15 UTC" "2018-05-15 14:50:18 UTC"
#> [423] "2018-05-15 14:54:17 UTC" "2018-05-15 14:58:18 UTC"
#> [425] "2018-05-15 15:02:19 UTC" "2018-05-15 15:06:21 UTC"
#> [427] "2018-05-15 15:10:23 UTC" "2018-05-15 15:14:24 UTC"
#> [429] "2018-05-15 15:18:26 UTC" "2018-05-15 15:22:25 UTC"
#> [431] "2018-05-15 15:26:28 UTC" "2018-05-15 15:30:29 UTC"
#> [433] "2018-05-15 15:34:31 UTC" "2018-05-15 15:38:33 UTC"
#> [435] "2018-05-15 15:42:33 UTC" "2018-05-15 15:46:35 UTC"
#> [437] "2018-05-15 15:50:37 UTC" "2018-05-15 15:54:37 UTC"
#> [439] "2018-05-15 15:58:38 UTC" "2018-05-15 16:02:38 UTC"
#> [441] "2018-05-15 16:06:41 UTC" "2018-05-15 16:10:40 UTC"
#> [443] "2018-05-15 16:14:41 UTC" "2018-05-15 16:18:45 UTC"
#> [445] "2018-05-15 16:22:45 UTC" "2018-05-15 16:26:45 UTC"
#> [447] "2018-05-15 16:30:47 UTC" "2018-05-15 16:34:47 UTC"
#> [449] "2018-05-15 16:38:50 UTC" "2018-05-15 16:42:49 UTC"
#> [451] "2018-05-15 16:46:51 UTC" "2018-05-15 16:50:54 UTC"
#> [453] "2018-05-15 16:54:53 UTC" "2018-05-15 16:58:56 UTC"
#> [455] "2018-05-15 17:02:56 UTC" "2018-05-15 17:06:56 UTC"
#> [457] "2018-05-15 17:11:00 UTC" "2018-05-15 17:15:00 UTC"
unique(mt_time_lags(move_data, units = "s"))
#> Units: [s]
#>   [1] 241.6083 241.2145 242.0282 241.1976 241.0800 240.5189 240.0864 242.6678
#>   [9] 242.2705 238.7727 242.9916 239.6901 243.0046 240.7830 240.3434 242.4782
#>  [17] 240.3580 240.8113 241.0018 241.7700 241.6667 239.4957 244.1606 240.6535
#>  [25] 241.1765 240.0880 242.5980 241.0853 244.1392 239.1328 240.2226 242.4814
#>  [33] 242.0443 239.6373 241.0358 241.2882 238.8239 244.0303 240.1596 241.8291
#>  [41] 241.3348 241.3921 240.8604 241.7494 241.1791 241.2913 241.1070 240.9777
#>  [49] 241.6984 241.5453 239.4714 241.0169 241.2283 242.1130 240.7176 243.8555
#>  [57] 240.2207 241.7677 241.4961 240.2522 242.0899 240.6328 241.2259 239.7097
#>  [65] 242.7299 240.4443 242.8381 241.0431 238.8728 241.9052 242.6477 240.3644
#>  [73] 241.0568 241.0865 241.1863 242.5088 242.2024 238.5473 241.2349 241.7509
#>  [81] 242.6181 241.3337 242.5799 240.4383 240.5052 241.0815 241.1311 243.6101
#>  [89] 239.0590 242.6740 238.2350 242.6652 240.8537 241.0975 240.7945 243.9699
#>  [97] 240.0049 240.7602 240.2584 243.1325 239.9895 242.3360 242.5562 239.5827
#> [105] 240.6223 241.9335 240.3939 243.1982 240.6255 241.1742 239.7726 242.1873
#> [113] 241.6624 241.6780 241.0878 242.3536 239.4923 240.2693 242.7713 240.7574
#> [121] 241.8579 241.2966 242.7186 240.1026 240.8366 238.6687 244.3361 240.3348
#> [129] 241.5254 241.9341 242.1811 238.5849 242.1992 240.7525 241.7235 241.7701
#> [137] 240.4626 240.9701 242.1671 240.3428 241.9302 239.6675 242.3009       NA
#> [145] 240.3284 241.7160 239.2920 242.0145 243.4952 244.1881 239.2910 241.2374
#> [153] 241.1230 241.7256 240.1026 244.2577 240.4923 241.7806 242.0237 242.7323
#> [161] 239.0670 243.0420 240.7175 242.7525 239.7340 240.7774 241.0071 241.5131
#> [169] 243.0937 244.2688 238.9673 240.7559 242.7426 240.6813 241.4997 241.9495
#> [177] 240.9063 243.5133 240.0627 240.5854 242.6872 240.9535 241.1872 241.6497
#> [185] 243.6831 238.7990 242.7737 241.5594 241.3162 239.5207 244.1235 240.6063
#> [193] 242.4657 241.4270 239.0707 241.4340 243.9848 241.4697 240.1422 243.3299
#> [201] 238.8353 242.4275 241.4527 241.2048 240.4635 244.2366 238.8262 243.0244
#> [209] 240.3073 243.8340 239.5003 242.7649 241.4716 242.0583 241.7236 240.6574
#> [217] 240.2322 242.7952 240.3016 241.2116 242.3906 241.8397 243.1165 240.3743
#> [225] 242.5443 240.5542 242.9370 237.8217 243.8437 241.7842 241.1153 239.7881
#> [233] 244.4294 240.1573 240.4429 240.8028 244.4266 237.4220 245.1558 239.4410
#> [241] 243.7862 241.3033 240.6132 240.4161 241.0621 244.1778 240.9851 239.1208
#> [249] 243.2708 239.7995 244.2150 239.2414 243.0662 240.2131 243.2093 241.6665
#> [257] 241.5501 241.2945 241.6370 240.4948 240.4259 244.3213 240.5839 240.1962
#> [265] 241.7041 242.1898 240.7271 241.3300 243.6073 241.6794 239.6966 241.1461
#> [273] 241.9004 243.4793 239.1218 243.5077 239.8575 242.8818 241.0846 242.5338
#> [281] 240.0215 242.3985 238.8235 244.0283 241.3453 241.4702 240.4271 241.4800
#> [289] 240.2739 242.9010 242.2551 239.9529 243.9245 242.6878 237.8560 241.9667
#> [297] 242.3514 241.8261 241.5315 241.1998 242.0510 241.1308 241.4077 240.4810
#> [305] 244.6335 237.8834 242.7556 239.9794 241.5448 240.0715 242.8206 240.5081
#> [313] 240.6471 242.8339 240.4542 242.8255 240.0429 240.9245 241.4166 238.7439
#> [321] 242.4522 239.7140 244.6859 239.4570 238.7810 243.2053 241.6590 242.0458
#> [329] 241.3231 240.7171 240.8494 241.8200 240.4360 241.1338 240.6458 243.3808
#> [337] 239.6746 240.2793 242.5811 240.6202 239.7726 242.3185 242.1974 239.9539
#> [345] 241.9238 240.9650 239.3941 243.0409 241.8621 239.0461 242.3036 242.4996
#> [353] 239.0054 243.3176 239.1111 242.2014 241.9988 240.0427 241.7212 240.4545
#> [361] 241.7256 240.2928 241.8852 241.2832 242.0152 239.6331 241.2769 241.5694
#> [369] 243.7453 237.4964 243.4669 239.6761 241.6981 238.4339 244.6769 239.1094
#> [377] 242.0774 240.8941 242.5137 240.1021 239.9684 241.4454 242.4545 242.7366
#> [385] 239.7685 240.8254 241.1121 241.6384 242.5626 240.0111 241.4732 240.1269
#> [393] 242.0252 238.6574 244.6887 238.6961 241.9646 239.8409 240.8040 243.3231
#> [401] 240.7964 240.5082 240.2386 242.8502 241.7415 241.0378 241.1028 242.4221
#> [409] 238.8342 242.3693 241.7228 239.5492 242.2003 238.8348 242.6354 242.3375
#> [417] 240.1639 239.6775 242.4836 242.4547 239.2575 241.0679 241.0640 241.7842
#> [425] 242.0069 240.8657 242.2152 238.6945 243.3354 240.6193 242.4493 242.1146
#> [433] 239.7830 241.3433 242.1991 239.7966 241.1279 240.0602 243.7447 238.5376
#> [441] 241.2397 243.4801 240.1120 239.9102 242.2425 240.0775 242.8596 238.9809
#> [449] 241.8319 242.8781 239.7313 242.6297 239.8637 240.3315 243.3408 240.9463

# use align_move to interpolate move_data to a uniform time scale and lag.
# e.g. setting a resolution of 4 minutes:
m <- align_move(m = move_data, res = units::set_units(4, "min"))
#> Temporal resolution of 4 [min] is used to align trajectories.
# check the result: records with attribute interpolate == TRUE have been added
# all trajectories are now aligend to a uniform 4-minute resolution:
unique(mt_time_lags(m, units = "min"))
#> Units: [min]
#> [1]  4 NA

# same with resolution of 1 hour:
m <- align_move(move_data, res = units::set_units(1, "hour"))
#> Temporal resolution of 1 [h] is used to align trajectories.
unique(mt_time_lags(m, units = "hour"))
#> Units: [h]
#> [1]  1 NA

# resolution of 15 seconds:
m <- align_move(move_data, res = units::set_units(15, "sec"))
#> Temporal resolution of 15 [s] is used to align trajectories.
unique(mt_time_lags(m, units = "sec"))
#> Units: [s]
#> [1] 15 NA

# you can set the start/end times if needed:
# first, let us retrieve the start and end times
start_end_time <- range(mt_time(m))
start_end_time
#> [1] "2018-05-15 07:00:14 UTC" "2018-05-15 19:34:44 UTC"

# I want the start time to be at 00 minutes and 00 seconds for the first track:
start_end_time <- round.POSIXt(start_end_time, units = "hours")

m <- align_move(
  move_data, res = units::set_units(4, "min"), 
  start_end_time = start_end_time
)
#> Temporal resolution of 4 [min] is used to align trajectories.
mt_time(m)
#>   [1] "2018-05-15 10:00:00 UTC" "2018-05-15 10:04:00 UTC"
#>   [3] "2018-05-15 10:08:00 UTC" "2018-05-15 10:12:00 UTC"
#>   [5] "2018-05-15 10:16:00 UTC" "2018-05-15 10:20:00 UTC"
#>   [7] "2018-05-15 10:24:00 UTC" "2018-05-15 10:28:00 UTC"
#>   [9] "2018-05-15 10:32:00 UTC" "2018-05-15 10:36:00 UTC"
#>  [11] "2018-05-15 10:40:00 UTC" "2018-05-15 10:44:00 UTC"
#>  [13] "2018-05-15 10:48:00 UTC" "2018-05-15 10:52:00 UTC"
#>  [15] "2018-05-15 10:56:00 UTC" "2018-05-15 11:00:00 UTC"
#>  [17] "2018-05-15 11:04:00 UTC" "2018-05-15 11:08:00 UTC"
#>  [19] "2018-05-15 11:12:00 UTC" "2018-05-15 11:16:00 UTC"
#>  [21] "2018-05-15 11:20:00 UTC" "2018-05-15 11:24:00 UTC"
#>  [23] "2018-05-15 11:28:00 UTC" "2018-05-15 11:32:00 UTC"
#>  [25] "2018-05-15 11:36:00 UTC" "2018-05-15 11:40:00 UTC"
#>  [27] "2018-05-15 11:44:00 UTC" "2018-05-15 11:48:00 UTC"
#>  [29] "2018-05-15 11:52:00 UTC" "2018-05-15 11:56:00 UTC"
#>  [31] "2018-05-15 12:00:00 UTC" "2018-05-15 12:04:00 UTC"
#>  [33] "2018-05-15 12:08:00 UTC" "2018-05-15 12:12:00 UTC"
#>  [35] "2018-05-15 12:16:00 UTC" "2018-05-15 12:20:00 UTC"
#>  [37] "2018-05-15 12:24:00 UTC" "2018-05-15 12:28:00 UTC"
#>  [39] "2018-05-15 12:32:00 UTC" "2018-05-15 12:36:00 UTC"
#>  [41] "2018-05-15 12:40:00 UTC" "2018-05-15 12:44:00 UTC"
#>  [43] "2018-05-15 12:48:00 UTC" "2018-05-15 12:52:00 UTC"
#>  [45] "2018-05-15 12:56:00 UTC" "2018-05-15 13:00:00 UTC"
#>  [47] "2018-05-15 13:04:00 UTC" "2018-05-15 13:08:00 UTC"
#>  [49] "2018-05-15 13:12:00 UTC" "2018-05-15 13:16:00 UTC"
#>  [51] "2018-05-15 13:20:00 UTC" "2018-05-15 13:24:00 UTC"
#>  [53] "2018-05-15 13:28:00 UTC" "2018-05-15 13:32:00 UTC"
#>  [55] "2018-05-15 13:36:00 UTC" "2018-05-15 13:40:00 UTC"
#>  [57] "2018-05-15 13:44:00 UTC" "2018-05-15 13:48:00 UTC"
#>  [59] "2018-05-15 13:52:00 UTC" "2018-05-15 13:56:00 UTC"
#>  [61] "2018-05-15 14:00:00 UTC" "2018-05-15 14:04:00 UTC"
#>  [63] "2018-05-15 14:08:00 UTC" "2018-05-15 14:12:00 UTC"
#>  [65] "2018-05-15 14:16:00 UTC" "2018-05-15 14:20:00 UTC"
#>  [67] "2018-05-15 14:24:00 UTC" "2018-05-15 14:28:00 UTC"
#>  [69] "2018-05-15 14:32:00 UTC" "2018-05-15 14:36:00 UTC"
#>  [71] "2018-05-15 14:40:00 UTC" "2018-05-15 14:44:00 UTC"
#>  [73] "2018-05-15 14:48:00 UTC" "2018-05-15 14:52:00 UTC"
#>  [75] "2018-05-15 14:56:00 UTC" "2018-05-15 15:00:00 UTC"
#>  [77] "2018-05-15 15:04:00 UTC" "2018-05-15 15:08:00 UTC"
#>  [79] "2018-05-15 15:12:00 UTC" "2018-05-15 15:16:00 UTC"
#>  [81] "2018-05-15 15:20:00 UTC" "2018-05-15 15:24:00 UTC"
#>  [83] "2018-05-15 15:28:00 UTC" "2018-05-15 15:32:00 UTC"
#>  [85] "2018-05-15 15:36:00 UTC" "2018-05-15 15:40:00 UTC"
#>  [87] "2018-05-15 15:44:00 UTC" "2018-05-15 15:48:00 UTC"
#>  [89] "2018-05-15 15:52:00 UTC" "2018-05-15 15:56:00 UTC"
#>  [91] "2018-05-15 16:00:00 UTC" "2018-05-15 16:04:00 UTC"
#>  [93] "2018-05-15 16:08:00 UTC" "2018-05-15 16:12:00 UTC"
#>  [95] "2018-05-15 16:16:00 UTC" "2018-05-15 16:20:00 UTC"
#>  [97] "2018-05-15 16:24:00 UTC" "2018-05-15 16:28:00 UTC"
#>  [99] "2018-05-15 16:32:00 UTC" "2018-05-15 16:36:00 UTC"
#> [101] "2018-05-15 16:40:00 UTC" "2018-05-15 16:44:00 UTC"
#> [103] "2018-05-15 16:48:00 UTC" "2018-05-15 16:52:00 UTC"
#> [105] "2018-05-15 16:56:00 UTC" "2018-05-15 17:00:00 UTC"
#> [107] "2018-05-15 17:04:00 UTC" "2018-05-15 17:08:00 UTC"
#> [109] "2018-05-15 17:12:00 UTC" "2018-05-15 17:16:00 UTC"
#> [111] "2018-05-15 17:20:00 UTC" "2018-05-15 17:24:00 UTC"
#> [113] "2018-05-15 17:28:00 UTC" "2018-05-15 17:32:00 UTC"
#> [115] "2018-05-15 17:36:00 UTC" "2018-05-15 17:40:00 UTC"
#> [117] "2018-05-15 17:44:00 UTC" "2018-05-15 17:48:00 UTC"
#> [119] "2018-05-15 17:52:00 UTC" "2018-05-15 17:56:00 UTC"
#> [121] "2018-05-15 18:00:00 UTC" "2018-05-15 18:04:00 UTC"
#> [123] "2018-05-15 18:08:00 UTC" "2018-05-15 18:12:00 UTC"
#> [125] "2018-05-15 18:16:00 UTC" "2018-05-15 18:20:00 UTC"
#> [127] "2018-05-15 18:24:00 UTC" "2018-05-15 18:28:00 UTC"
#> [129] "2018-05-15 18:32:00 UTC" "2018-05-15 18:36:00 UTC"
#> [131] "2018-05-15 18:40:00 UTC" "2018-05-15 18:44:00 UTC"
#> [133] "2018-05-15 18:48:00 UTC" "2018-05-15 18:52:00 UTC"
#> [135] "2018-05-15 18:56:00 UTC" "2018-05-15 19:00:00 UTC"
#> [137] "2018-05-15 19:04:00 UTC" "2018-05-15 19:08:00 UTC"
#> [139] "2018-05-15 19:12:00 UTC" "2018-05-15 19:16:00 UTC"
#> [141] "2018-05-15 19:20:00 UTC" "2018-05-15 19:24:00 UTC"
#> [143] "2018-05-15 19:28:00 UTC" "2018-05-15 19:32:00 UTC"
#> [145] "2018-05-15 08:04:00 UTC" "2018-05-15 08:08:00 UTC"
#> [147] "2018-05-15 08:12:00 UTC" "2018-05-15 08:16:00 UTC"
#> [149] "2018-05-15 08:20:00 UTC" "2018-05-15 08:24:00 UTC"
#> [151] "2018-05-15 08:28:00 UTC" "2018-05-15 08:32:00 UTC"
#> [153] "2018-05-15 08:36:00 UTC" "2018-05-15 08:40:00 UTC"
#> [155] "2018-05-15 08:44:00 UTC" "2018-05-15 08:48:00 UTC"
#> [157] "2018-05-15 08:52:00 UTC" "2018-05-15 08:56:00 UTC"
#> [159] "2018-05-15 09:00:00 UTC" "2018-05-15 09:04:00 UTC"
#> [161] "2018-05-15 09:08:00 UTC" "2018-05-15 09:12:00 UTC"
#> [163] "2018-05-15 09:16:00 UTC" "2018-05-15 09:20:00 UTC"
#> [165] "2018-05-15 09:24:00 UTC" "2018-05-15 09:28:00 UTC"
#> [167] "2018-05-15 09:32:00 UTC" "2018-05-15 09:36:00 UTC"
#> [169] "2018-05-15 09:40:00 UTC" "2018-05-15 09:44:00 UTC"
#> [171] "2018-05-15 09:48:00 UTC" "2018-05-15 09:52:00 UTC"
#> [173] "2018-05-15 09:56:00 UTC" "2018-05-15 10:00:00 UTC"
#> [175] "2018-05-15 10:04:00 UTC" "2018-05-15 10:08:00 UTC"
#> [177] "2018-05-15 10:12:00 UTC" "2018-05-15 10:16:00 UTC"
#> [179] "2018-05-15 10:20:00 UTC" "2018-05-15 10:24:00 UTC"
#> [181] "2018-05-15 10:28:00 UTC" "2018-05-15 10:32:00 UTC"
#> [183] "2018-05-15 10:36:00 UTC" "2018-05-15 10:40:00 UTC"
#> [185] "2018-05-15 10:44:00 UTC" "2018-05-15 10:48:00 UTC"
#> [187] "2018-05-15 10:52:00 UTC" "2018-05-15 10:56:00 UTC"
#> [189] "2018-05-15 11:00:00 UTC" "2018-05-15 11:04:00 UTC"
#> [191] "2018-05-15 11:08:00 UTC" "2018-05-15 11:12:00 UTC"
#> [193] "2018-05-15 11:16:00 UTC" "2018-05-15 11:20:00 UTC"
#> [195] "2018-05-15 11:24:00 UTC" "2018-05-15 11:28:00 UTC"
#> [197] "2018-05-15 11:32:00 UTC" "2018-05-15 11:36:00 UTC"
#> [199] "2018-05-15 11:40:00 UTC" "2018-05-15 11:44:00 UTC"
#> [201] "2018-05-15 11:48:00 UTC" "2018-05-15 11:52:00 UTC"
#> [203] "2018-05-15 11:56:00 UTC" "2018-05-15 12:00:00 UTC"
#> [205] "2018-05-15 12:04:00 UTC" "2018-05-15 12:08:00 UTC"
#> [207] "2018-05-15 12:12:00 UTC" "2018-05-15 12:16:00 UTC"
#> [209] "2018-05-15 12:20:00 UTC" "2018-05-15 12:24:00 UTC"
#> [211] "2018-05-15 12:28:00 UTC" "2018-05-15 12:32:00 UTC"
#> [213] "2018-05-15 12:36:00 UTC" "2018-05-15 12:40:00 UTC"
#> [215] "2018-05-15 12:44:00 UTC" "2018-05-15 12:48:00 UTC"
#> [217] "2018-05-15 12:52:00 UTC" "2018-05-15 12:56:00 UTC"
#> [219] "2018-05-15 13:00:00 UTC" "2018-05-15 13:04:00 UTC"
#> [221] "2018-05-15 13:08:00 UTC" "2018-05-15 13:12:00 UTC"
#> [223] "2018-05-15 13:16:00 UTC" "2018-05-15 13:20:00 UTC"
#> [225] "2018-05-15 13:24:00 UTC" "2018-05-15 13:28:00 UTC"
#> [227] "2018-05-15 13:32:00 UTC" "2018-05-15 13:36:00 UTC"
#> [229] "2018-05-15 13:40:00 UTC" "2018-05-15 13:44:00 UTC"
#> [231] "2018-05-15 13:48:00 UTC" "2018-05-15 13:52:00 UTC"
#> [233] "2018-05-15 13:56:00 UTC" "2018-05-15 14:00:00 UTC"
#> [235] "2018-05-15 14:04:00 UTC" "2018-05-15 14:08:00 UTC"
#> [237] "2018-05-15 14:12:00 UTC" "2018-05-15 14:16:00 UTC"
#> [239] "2018-05-15 14:20:00 UTC" "2018-05-15 14:24:00 UTC"
#> [241] "2018-05-15 14:28:00 UTC" "2018-05-15 14:32:00 UTC"
#> [243] "2018-05-15 14:36:00 UTC" "2018-05-15 14:40:00 UTC"
#> [245] "2018-05-15 14:44:00 UTC" "2018-05-15 14:48:00 UTC"
#> [247] "2018-05-15 14:52:00 UTC" "2018-05-15 14:56:00 UTC"
#> [249] "2018-05-15 15:00:00 UTC" "2018-05-15 15:04:00 UTC"
#> [251] "2018-05-15 15:08:00 UTC" "2018-05-15 15:12:00 UTC"
#> [253] "2018-05-15 15:16:00 UTC" "2018-05-15 15:20:00 UTC"
#> [255] "2018-05-15 15:24:00 UTC" "2018-05-15 15:28:00 UTC"
#> [257] "2018-05-15 15:32:00 UTC" "2018-05-15 15:36:00 UTC"
#> [259] "2018-05-15 15:40:00 UTC" "2018-05-15 15:44:00 UTC"
#> [261] "2018-05-15 15:48:00 UTC" "2018-05-15 15:52:00 UTC"
#> [263] "2018-05-15 15:56:00 UTC" "2018-05-15 16:00:00 UTC"
#> [265] "2018-05-15 16:04:00 UTC" "2018-05-15 16:08:00 UTC"
#> [267] "2018-05-15 16:12:00 UTC" "2018-05-15 16:16:00 UTC"
#> [269] "2018-05-15 16:20:00 UTC" "2018-05-15 16:24:00 UTC"
#> [271] "2018-05-15 16:28:00 UTC" "2018-05-15 16:32:00 UTC"
#> [273] "2018-05-15 16:36:00 UTC" "2018-05-15 16:40:00 UTC"
#> [275] "2018-05-15 16:44:00 UTC" "2018-05-15 16:48:00 UTC"
#> [277] "2018-05-15 16:52:00 UTC" "2018-05-15 16:56:00 UTC"
#> [279] "2018-05-15 17:00:00 UTC" "2018-05-15 17:04:00 UTC"
#> [281] "2018-05-15 17:08:00 UTC" "2018-05-15 17:12:00 UTC"
#> [283] "2018-05-15 17:16:00 UTC" "2018-05-15 17:20:00 UTC"
#> [285] "2018-05-15 17:24:00 UTC" "2018-05-15 17:28:00 UTC"
#> [287] "2018-05-15 17:32:00 UTC" "2018-05-15 17:36:00 UTC"
#> [289] "2018-05-15 17:40:00 UTC" "2018-05-15 17:44:00 UTC"
#> [291] "2018-05-15 17:48:00 UTC" "2018-05-15 17:52:00 UTC"
#> [293] "2018-05-15 17:56:00 UTC" "2018-05-15 18:00:00 UTC"
#> [295] "2018-05-15 18:04:00 UTC" "2018-05-15 18:08:00 UTC"
#> [297] "2018-05-15 18:12:00 UTC" "2018-05-15 18:16:00 UTC"
#> [299] "2018-05-15 18:20:00 UTC" "2018-05-15 18:24:00 UTC"
#> [301] "2018-05-15 18:28:00 UTC" "2018-05-15 18:32:00 UTC"
#> [303] "2018-05-15 18:36:00 UTC" "2018-05-15 18:40:00 UTC"
#> [305] "2018-05-15 07:00:00 UTC" "2018-05-15 07:04:00 UTC"
#> [307] "2018-05-15 07:08:00 UTC" "2018-05-15 07:12:00 UTC"
#> [309] "2018-05-15 07:16:00 UTC" "2018-05-15 07:20:00 UTC"
#> [311] "2018-05-15 07:24:00 UTC" "2018-05-15 07:28:00 UTC"
#> [313] "2018-05-15 07:32:00 UTC" "2018-05-15 07:36:00 UTC"
#> [315] "2018-05-15 07:40:00 UTC" "2018-05-15 07:44:00 UTC"
#> [317] "2018-05-15 07:48:00 UTC" "2018-05-15 07:52:00 UTC"
#> [319] "2018-05-15 07:56:00 UTC" "2018-05-15 08:00:00 UTC"
#> [321] "2018-05-15 08:04:00 UTC" "2018-05-15 08:08:00 UTC"
#> [323] "2018-05-15 08:12:00 UTC" "2018-05-15 08:16:00 UTC"
#> [325] "2018-05-15 08:20:00 UTC" "2018-05-15 08:24:00 UTC"
#> [327] "2018-05-15 08:28:00 UTC" "2018-05-15 08:32:00 UTC"
#> [329] "2018-05-15 08:36:00 UTC" "2018-05-15 08:40:00 UTC"
#> [331] "2018-05-15 08:44:00 UTC" "2018-05-15 08:48:00 UTC"
#> [333] "2018-05-15 08:52:00 UTC" "2018-05-15 08:56:00 UTC"
#> [335] "2018-05-15 09:00:00 UTC" "2018-05-15 09:04:00 UTC"
#> [337] "2018-05-15 09:08:00 UTC" "2018-05-15 09:12:00 UTC"
#> [339] "2018-05-15 09:16:00 UTC" "2018-05-15 09:20:00 UTC"
#> [341] "2018-05-15 09:24:00 UTC" "2018-05-15 09:28:00 UTC"
#> [343] "2018-05-15 09:32:00 UTC" "2018-05-15 09:36:00 UTC"
#> [345] "2018-05-15 09:40:00 UTC" "2018-05-15 09:44:00 UTC"
#> [347] "2018-05-15 09:48:00 UTC" "2018-05-15 09:52:00 UTC"
#> [349] "2018-05-15 09:56:00 UTC" "2018-05-15 10:00:00 UTC"
#> [351] "2018-05-15 10:04:00 UTC" "2018-05-15 10:08:00 UTC"
#> [353] "2018-05-15 10:12:00 UTC" "2018-05-15 10:16:00 UTC"
#> [355] "2018-05-15 10:20:00 UTC" "2018-05-15 10:24:00 UTC"
#> [357] "2018-05-15 10:28:00 UTC" "2018-05-15 10:32:00 UTC"
#> [359] "2018-05-15 10:36:00 UTC" "2018-05-15 10:40:00 UTC"
#> [361] "2018-05-15 10:44:00 UTC" "2018-05-15 10:48:00 UTC"
#> [363] "2018-05-15 10:52:00 UTC" "2018-05-15 10:56:00 UTC"
#> [365] "2018-05-15 11:00:00 UTC" "2018-05-15 11:04:00 UTC"
#> [367] "2018-05-15 11:08:00 UTC" "2018-05-15 11:12:00 UTC"
#> [369] "2018-05-15 11:16:00 UTC" "2018-05-15 11:20:00 UTC"
#> [371] "2018-05-15 11:24:00 UTC" "2018-05-15 11:28:00 UTC"
#> [373] "2018-05-15 11:32:00 UTC" "2018-05-15 11:36:00 UTC"
#> [375] "2018-05-15 11:40:00 UTC" "2018-05-15 11:44:00 UTC"
#> [377] "2018-05-15 11:48:00 UTC" "2018-05-15 11:52:00 UTC"
#> [379] "2018-05-15 11:56:00 UTC" "2018-05-15 12:00:00 UTC"
#> [381] "2018-05-15 12:04:00 UTC" "2018-05-15 12:08:00 UTC"
#> [383] "2018-05-15 12:12:00 UTC" "2018-05-15 12:16:00 UTC"
#> [385] "2018-05-15 12:20:00 UTC" "2018-05-15 12:24:00 UTC"
#> [387] "2018-05-15 12:28:00 UTC" "2018-05-15 12:32:00 UTC"
#> [389] "2018-05-15 12:36:00 UTC" "2018-05-15 12:40:00 UTC"
#> [391] "2018-05-15 12:44:00 UTC" "2018-05-15 12:48:00 UTC"
#> [393] "2018-05-15 12:52:00 UTC" "2018-05-15 12:56:00 UTC"
#> [395] "2018-05-15 13:00:00 UTC" "2018-05-15 13:04:00 UTC"
#> [397] "2018-05-15 13:08:00 UTC" "2018-05-15 13:12:00 UTC"
#> [399] "2018-05-15 13:16:00 UTC" "2018-05-15 13:20:00 UTC"
#> [401] "2018-05-15 13:24:00 UTC" "2018-05-15 13:28:00 UTC"
#> [403] "2018-05-15 13:32:00 UTC" "2018-05-15 13:36:00 UTC"
#> [405] "2018-05-15 13:40:00 UTC" "2018-05-15 13:44:00 UTC"
#> [407] "2018-05-15 13:48:00 UTC" "2018-05-15 13:52:00 UTC"
#> [409] "2018-05-15 13:56:00 UTC" "2018-05-15 14:00:00 UTC"
#> [411] "2018-05-15 14:04:00 UTC" "2018-05-15 14:08:00 UTC"
#> [413] "2018-05-15 14:12:00 UTC" "2018-05-15 14:16:00 UTC"
#> [415] "2018-05-15 14:20:00 UTC" "2018-05-15 14:24:00 UTC"
#> [417] "2018-05-15 14:28:00 UTC" "2018-05-15 14:32:00 UTC"
#> [419] "2018-05-15 14:36:00 UTC" "2018-05-15 14:40:00 UTC"
#> [421] "2018-05-15 14:44:00 UTC" "2018-05-15 14:48:00 UTC"
#> [423] "2018-05-15 14:52:00 UTC" "2018-05-15 14:56:00 UTC"
#> [425] "2018-05-15 15:00:00 UTC" "2018-05-15 15:04:00 UTC"
#> [427] "2018-05-15 15:08:00 UTC" "2018-05-15 15:12:00 UTC"
#> [429] "2018-05-15 15:16:00 UTC" "2018-05-15 15:20:00 UTC"
#> [431] "2018-05-15 15:24:00 UTC" "2018-05-15 15:28:00 UTC"
#> [433] "2018-05-15 15:32:00 UTC" "2018-05-15 15:36:00 UTC"
#> [435] "2018-05-15 15:40:00 UTC" "2018-05-15 15:44:00 UTC"
#> [437] "2018-05-15 15:48:00 UTC" "2018-05-15 15:52:00 UTC"
#> [439] "2018-05-15 15:56:00 UTC" "2018-05-15 16:00:00 UTC"
#> [441] "2018-05-15 16:04:00 UTC" "2018-05-15 16:08:00 UTC"
#> [443] "2018-05-15 16:12:00 UTC" "2018-05-15 16:16:00 UTC"
#> [445] "2018-05-15 16:20:00 UTC" "2018-05-15 16:24:00 UTC"
#> [447] "2018-05-15 16:28:00 UTC" "2018-05-15 16:32:00 UTC"
#> [449] "2018-05-15 16:36:00 UTC" "2018-05-15 16:40:00 UTC"
#> [451] "2018-05-15 16:44:00 UTC" "2018-05-15 16:48:00 UTC"
#> [453] "2018-05-15 16:52:00 UTC" "2018-05-15 16:56:00 UTC"
#> [455] "2018-05-15 17:00:00 UTC" "2018-05-15 17:04:00 UTC"
#> [457] "2018-05-15 17:08:00 UTC" "2018-05-15 17:12:00 UTC"