Example NDVI data

This dataset is a SpatRasterDataset, representing simulated NDVI images covering the Lake of Constance area, as well as invented dates and times that simulate acquisition times.

Usage

example_data(file = "basemap_data.rds")

Format

SpatRasterDataset, consisting of NDVI layers and acquisition times as POSIXct objects.

Source

Simulated based on MODIS (MOD13Q1 NDVI)

Arguments

file

character, file name of the example dataset.

Details

This object is used by some moveVis examples and unit tests. Use readRDS(example_data()) to retrieve the SpatRasterDataset.

Times can be retrieved using terra::time(readRDS(example_data()))

Note

All data contained should only be used for testing moveVis and are not suitable to be used for analysis or interpretation.