Convert a data.frame into a move or moveStack object

This function is a simple wrapper that converts a data.frame into a move or moveStack object. Both can be used as inputs to frames_spatial or frames_graph.

df2move(df, proj, x, y, time, track_id = NULL, data = NULL, ...)

Arguments

df

data.frame, a data.frame with rows representing observations and columns representing x and y coordinates, time and optionally track IDs, if multiple tracks are contained.

proj

projection, character (proj4string) or CRS object, indicating the projection that the coordinates of df represent.

x

character, name of the column in df that represents x coordinates.

y

character, name of the column in df that represents y coordinates.

time

character, name of the column in df that represents timestamps. Timestamps need to be of class POSIXct.

track_id

character, optional, name of the column in df that represents track names or IDs. If set, a moveStack is returned, otherwise, a move object is returned.

data

data.frame, optional, to add additional data such as path colours (see move). Number of rows must equal number of rows of df.

...

additional arguments passed to move.

Value

A move or moveStack object.

See also

Examples

library(moveVis)
library(move)

# load the example data and convert them into a data.frame
data("move_data")
move_df <- methods::as(move_data, "data.frame")

# use df2move to convert the data.frame into a moveStack
df2move(move_df,
        proj = "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0", 
        x = "coords.x1", y = "coords.x2", time = "timestamps", track_id = "trackId")
#> Warning: GDAL Message 1: +init=epsg:XXXX syntax is deprecated. It might return a CRS with a non-EPSG compliant axis order.
#> Warning: Discarded datum Unknown based on WGS84 ellipsoid in Proj4 definition
#>              coordinates        x        y                       time
#> 1   (8.964786, 47.75646) 8.964786 47.75646 2018-05-15 09:59:59.856108
#> 2   (8.964773, 47.75667) 8.964773 47.75667 2018-05-15 10:04:01.464359
#> 3   (8.964713, 47.75683) 8.964713 47.75683 2018-05-15 10:08:02.678875
#> 4   (8.964644, 47.75695) 8.964644 47.75695 2018-05-15 10:12:04.707085
#> 5   (8.964526, 47.75706) 8.964526 47.75706 2018-05-15 10:16:05.904692
#> 6    (8.96425, 47.75724) 8.964250 47.75724 2018-05-15 10:20:06.984729
#> 7   (8.963802, 47.75722) 8.963802 47.75722 2018-05-15 10:24:07.503590
#> 8   (8.963486, 47.75724) 8.963486 47.75724 2018-05-15 10:28:07.589973
#> 9   (8.963229, 47.75736) 8.963229 47.75736 2018-05-15 10:32:10.257785
#> 10   (8.962975, 47.7575) 8.962975 47.75750 2018-05-15 10:36:12.528266
#> 11   (8.96269, 47.75754) 8.962690 47.75754 2018-05-15 10:40:11.300930
#> 12  (8.962433, 47.75753) 8.962433 47.75753 2018-05-15 10:44:14.292572
#> 13  (8.962028, 47.75751) 8.962028 47.75751 2018-05-15 10:48:13.982639
#> 14   (8.961616, 47.7575) 8.961616 47.75750 2018-05-15 10:52:16.987274
#> 15   (8.961182, 47.7575) 8.961182 47.75750 2018-05-15 10:56:17.770314
#> 16  (8.960484, 47.75756) 8.960484 47.75756 2018-05-15 11:00:18.113693
#> 17  (8.957716, 47.75786) 8.957716 47.75786 2018-05-15 11:04:20.591893
#> 18  (8.954149, 47.75836) 8.954149 47.75836 2018-05-15 11:08:20.949864
#> 19  (8.950258, 47.75945) 8.950258 47.75945 2018-05-15 11:12:21.761121
#> 20  (8.947356, 47.76117) 8.947356 47.76117 2018-05-15 11:16:22.762903
#> 21  (8.946179, 47.76336) 8.946179 47.76336 2018-05-15 11:20:24.532878
#> 22  (8.942071, 47.76589) 8.942071 47.76589 2018-05-15 11:24:26.199534
#> 23  (8.938741, 47.76757) 8.938741 47.76757 2018-05-15 11:28:25.695186
#> 24  (8.935022, 47.76816) 8.935022 47.76816 2018-05-15 11:32:29.855754
#> 25  (8.931244, 47.76792) 8.931244 47.76792 2018-05-15 11:36:30.509280
#> 26  (8.927847, 47.76718) 8.927847 47.76718 2018-05-15 11:40:31.685769
#> 27  (8.924086, 47.76587) 8.924086 47.76587 2018-05-15 11:44:31.773756
#> 28  (8.923044, 47.76406) 8.923044 47.76406 2018-05-15 11:48:34.371707
#> 29  (8.921451, 47.76273) 8.921451 47.76273 2018-05-15 11:52:35.457000
#> 30  (8.918889, 47.76249) 8.918889 47.76249 2018-05-15 11:56:39.596250
#> 31  (8.915441, 47.76317) 8.915441 47.76317 2018-05-15 12:00:38.729068
#> 32  (8.911054, 47.76246) 8.911054 47.76246 2018-05-15 12:04:38.951704
#> 33  (8.910482, 47.75924) 8.910482 47.75924 2018-05-15 12:08:41.433076
#> 34  (8.908828, 47.75784) 8.908828 47.75784 2018-05-15 12:12:43.477401
#> 35  (8.906905, 47.75791) 8.906905 47.75791 2018-05-15 12:16:43.114655
#> 36  (8.905824, 47.75861) 8.905824 47.75861 2018-05-15 12:20:44.150446
#> 37   (8.905824, 47.7595) 8.905824 47.75950 2018-05-15 12:24:45.438684
#> 38  (8.906543, 47.76021) 8.906543 47.76021 2018-05-15 12:28:44.262552
#> 39  (8.906389, 47.76088) 8.906389 47.76088 2018-05-15 12:32:48.292811
#> 40  (8.906115, 47.76153) 8.906115 47.76153 2018-05-15 12:36:48.452394
#> 41  (8.906359, 47.76212) 8.906359 47.76212 2018-05-15 12:40:50.281519
#> 42  (8.907264, 47.76245) 8.907264 47.76245 2018-05-15 12:44:51.616330
#> 43  (8.908438, 47.76205) 8.908438 47.76205 2018-05-15 12:48:53.008382
#> 44    (8.90912, 47.7614) 8.909120 47.76140 2018-05-15 12:52:53.868752
#> 45   (8.90964, 47.76065) 8.909640 47.76065 2018-05-15 12:56:55.618126
#> 46   (8.90987, 47.75999) 8.909870 47.75999 2018-05-15 13:00:56.797220
#> 47  (8.909035, 47.75961) 8.909035 47.75961 2018-05-15 13:04:58.088518
#> 48  (8.908061, 47.75914) 8.908061 47.75914 2018-05-15 13:08:59.195510
#> 49   (8.907086, 47.7585) 8.907086 47.75850 2018-05-15 13:13:00.173186
#> 50  (8.906392, 47.75778) 8.906392 47.75778 2018-05-15 13:17:01.871578
#> 51   (8.90612, 47.75704) 8.906120 47.75704 2018-05-15 13:21:03.416925
#> 52  (8.906264, 47.75634) 8.906264 47.75634 2018-05-15 13:25:02.888335
#> 53  (8.906682, 47.75568) 8.906682 47.75568 2018-05-15 13:29:03.905260
#> 54  (8.907854, 47.75517) 8.907854 47.75517 2018-05-15 13:33:05.133542
#> 55  (8.910687, 47.75421) 8.910687 47.75421 2018-05-15 13:37:07.246579
#> 56  (8.914017, 47.75259) 8.914017 47.75259 2018-05-15 13:41:07.964181
#> 57  (8.915397, 47.75044) 8.915397 47.75044 2018-05-15 13:45:11.819677
#> 58  (8.916511, 47.74848) 8.916511 47.74848 2018-05-15 13:49:12.040368
#> 59  (8.918294, 47.74621) 8.918294 47.74621 2018-05-15 13:53:13.808099
#> 60   (8.920321, 47.7441) 8.920321 47.74410 2018-05-15 13:57:15.304163
#> 61  (8.922759, 47.74234) 8.922759 47.74234 2018-05-15 14:01:15.556404
#> 62   (8.925967, 47.7409) 8.925967 47.74090 2018-05-15 14:05:17.646316
#> 63   (8.92924, 47.73898) 8.929240 47.73898 2018-05-15 14:09:18.279098
#> 64  (8.932571, 47.73714) 8.932571 47.73714 2018-05-15 14:13:19.504998
#> 65   (8.935751, 47.7357) 8.935751 47.73570 2018-05-15 14:17:19.214716
#> 66  (8.939098, 47.73445) 8.939098 47.73445 2018-05-15 14:21:21.944616
#> 67  (8.942825, 47.73307) 8.942825 47.73307 2018-05-15 14:25:22.388902
#> 68  (8.944914, 47.73199) 8.944914 47.73199 2018-05-15 14:29:25.226975
#> 69  (8.949198, 47.73112) 8.949198 47.73112 2018-05-15 14:33:26.270066
#> 70  (8.954076, 47.73056) 8.954076 47.73056 2018-05-15 14:37:25.142850
#> 71  (8.958403, 47.72986) 8.958403 47.72986 2018-05-15 14:41:27.048003
#> 72  (8.961477, 47.72865) 8.961477 47.72865 2018-05-15 14:45:29.695678
#> 73   (8.964123, 47.7278) 8.964123 47.72780 2018-05-15 14:49:30.060102
#> 74  (8.968237, 47.72762) 8.968237 47.72762 2018-05-15 14:53:31.116892
#> 75  (8.972024, 47.72803) 8.972024 47.72803 2018-05-15 14:57:32.203414
#> 76  (8.974934, 47.72893) 8.974934 47.72893 2018-05-15 15:01:33.389695
#> 77  (8.976675, 47.73027) 8.976675 47.73027 2018-05-15 15:05:35.898451
#> 78  (8.976815, 47.73228) 8.976815 47.73228 2018-05-15 15:09:38.100896
#> 79  (8.976135, 47.73345) 8.976135 47.73345 2018-05-15 15:13:36.648150
#> 80  (8.975277, 47.73317) 8.975277 47.73317 2018-05-15 15:17:37.883014
#> 81   (8.97475, 47.73216) 8.974750 47.73216 2018-05-15 15:21:39.633903
#> 82  (8.974556, 47.73125) 8.974556 47.73125 2018-05-15 15:25:42.252052
#> 83  (8.973442, 47.73087) 8.973442 47.73087 2018-05-15 15:29:43.585767
#> 84  (8.971809, 47.73093) 8.971809 47.73093 2018-05-15 15:33:46.165633
#> 85  (8.970808, 47.73151) 8.970808 47.73151 2018-05-15 15:37:46.603945
#> 86  (8.970359, 47.73257) 8.970359 47.73257 2018-05-15 15:41:47.109115
#> 87  (8.970233, 47.73374) 8.970233 47.73374 2018-05-15 15:45:48.190608
#> 88  (8.970164, 47.73392) 8.970164 47.73392 2018-05-15 15:49:49.321668
#> 89  (8.970027, 47.73393) 8.970027 47.73393 2018-05-15 15:53:52.931804
#> 90  (8.969938, 47.73397) 8.969938 47.73397 2018-05-15 15:57:51.990802
#> 91  (8.969901, 47.73399) 8.969901 47.73399 2018-05-15 16:01:54.664852
#> 92  (8.969948, 47.73401) 8.969948 47.73401 2018-05-15 16:05:52.899834
#> 93   (8.970157, 47.7341) 8.970157 47.73410 2018-05-15 16:09:55.565039
#> 94  (8.970571, 47.73401) 8.970571 47.73401 2018-05-15 16:13:56.418772
#> 95  (8.970783, 47.73479) 8.970783 47.73479 2018-05-15 16:17:57.516311
#> 96   (8.971475, 47.7361) 8.971475 47.73610 2018-05-15 16:21:58.310763
#> 97  (8.972465, 47.73749) 8.972465 47.73749 2018-05-15 16:26:02.280614
#> 98  (8.972677, 47.73842) 8.972677 47.73842 2018-05-15 16:30:02.285493
#> 99  (8.973304, 47.73912) 8.973304 47.73912 2018-05-15 16:34:03.045708
#> 100 (8.972492, 47.73958) 8.972492 47.73958 2018-05-15 16:38:03.304062
#> 101  (8.971129, 47.7399) 8.971129 47.73990 2018-05-15 16:42:06.436585
#> 102 (8.969977, 47.74029) 8.969977 47.74029 2018-05-15 16:46:06.426115
#> 103 (8.969702, 47.74092) 8.969702 47.74092 2018-05-15 16:50:08.762083
#> 104 (8.971442, 47.74097) 8.971442 47.74097 2018-05-15 16:54:11.318312
#> 105 (8.975057, 47.74146) 8.975057 47.74146 2018-05-15 16:58:10.900989
#> 106 (8.978936, 47.74168) 8.978936 47.74168 2018-05-15 17:02:11.523268
#> 107 (8.982373, 47.74188) 8.982373 47.74188 2018-05-15 17:06:13.456778
#> 108 (8.985296, 47.74274) 8.985296 47.74274 2018-05-15 17:10:13.850667
#> 109 (8.989265, 47.74401) 8.989265 47.74401 2018-05-15 17:14:17.048910
#> 110 (8.988339, 47.74485) 8.988339 47.74485 2018-05-15 17:18:17.674374
#> 111 (8.988398, 47.74619) 8.988398 47.74619 2018-05-15 17:22:18.848532
#> 112 (8.990068, 47.74762) 8.990068 47.74762 2018-05-15 17:26:18.621167
#> 113 (8.991679, 47.74839) 8.991679 47.74839 2018-05-15 17:30:20.808508
#> 114 (8.991192, 47.74999) 8.991192 47.74999 2018-05-15 17:34:22.470931
#> 115 (8.989001, 47.75067) 8.989001 47.75067 2018-05-15 17:38:24.148919
#> 116 (8.987037, 47.75141) 8.987037 47.75141 2018-05-15 17:42:25.236712
#> 117   (8.98618, 47.7524) 8.986180 47.75240 2018-05-15 17:46:27.590352
#> 118 (8.986676, 47.75357) 8.986676 47.75357 2018-05-15 17:50:27.082609
#> 119 (8.987233, 47.75474) 8.987233 47.75474 2018-05-15 17:54:27.351872
#> 120 (8.985186, 47.75546) 8.985186 47.75546 2018-05-15 17:58:30.123211
#> 121 (8.983228, 47.75641) 8.983228 47.75641 2018-05-15 18:02:30.880573
#> 122 (8.981131, 47.75707) 8.981131 47.75707 2018-05-15 18:06:32.738479
#> 123   (8.9785, 47.75696) 8.978500 47.75696 2018-05-15 18:10:34.035042
#> 124 (8.976829, 47.75678) 8.976829 47.75678 2018-05-15 18:14:36.753658
#> 125 (8.974597, 47.75655) 8.974597 47.75655 2018-05-15 18:18:36.856305
#> 126 (8.972934, 47.75598) 8.972934 47.75598 2018-05-15 18:22:37.692944
#> 127 (8.971754, 47.75504) 8.971754 47.75504 2018-05-15 18:26:36.361680
#> 128 (8.970733, 47.75384) 8.970733 47.75384 2018-05-15 18:30:40.697784
#> 129 (8.969479, 47.75277) 8.969479 47.75277 2018-05-15 18:34:41.032557
#> 130 (8.967456, 47.75156) 8.967456 47.75156 2018-05-15 18:38:42.557943
#> 131 (8.965744, 47.75088) 8.965744 47.75088 2018-05-15 18:42:44.492061
#> 132 (8.964545, 47.75113) 8.964545 47.75113 2018-05-15 18:46:46.673126
#> 133 (8.964019, 47.75234) 8.964019 47.75234 2018-05-15 18:50:45.257986
#> 134 (8.964507, 47.75388) 8.964507 47.75388 2018-05-15 18:54:47.457188
#> 135 (8.965588, 47.75472) 8.965588 47.75472 2018-05-15 18:58:48.209663
#> 136 (8.965378, 47.75522) 8.965378 47.75522 2018-05-15 19:02:49.933120
#> 137  (8.964869, 47.7555) 8.964869 47.75550 2018-05-15 19:06:51.703185
#> 138 (8.964811, 47.75573) 8.964811 47.75573 2018-05-15 19:10:52.165816
#> 139 (8.965299, 47.75611) 8.965299 47.75611 2018-05-15 19:14:53.135956
#> 140 (8.965311, 47.75616) 8.965311 47.75616 2018-05-15 19:18:55.303095
#> 141 (8.965311, 47.75617) 8.965311 47.75617 2018-05-15 19:22:55.645930
#> 142 (8.964871, 47.75617) 8.964871 47.75617 2018-05-15 19:26:57.576142
#> 143 (8.964152, 47.75608) 8.964152 47.75608 2018-05-15 19:30:57.243616
#> 144 (8.964013, 47.75571) 8.964013 47.75571 2018-05-15 19:34:59.544519
#> 145 (8.919686, 47.73544) 8.919686 47.73544 2018-05-15 08:00:00.300123
#> 146 (8.919106, 47.73532) 8.919106 47.73532 2018-05-15 08:04:00.628543
#> 147 (8.918663, 47.73545) 8.918663 47.73545 2018-05-15 08:08:02.344581
#> 148 (8.918469, 47.73556) 8.918469 47.73556 2018-05-15 08:12:01.636576
#> 149  (8.91839, 47.73564) 8.918390 47.73564 2018-05-15 08:16:03.651039
#> 150 (8.918088, 47.73593) 8.918088 47.73593 2018-05-15 08:20:07.146231
#> 151 (8.917435, 47.73607) 8.917435 47.73607 2018-05-15 08:24:11.334365
#> 152  (8.916868, 47.7363) 8.916868 47.73630 2018-05-15 08:28:10.625339
#> 153 (8.916582, 47.73657) 8.916582 47.73657 2018-05-15 08:32:11.862788
#> 154 (8.916697, 47.73679) 8.916697 47.73679 2018-05-15 08:36:12.985825
#> 155 (8.917246, 47.73685) 8.917246 47.73685 2018-05-15 08:40:14.711384
#> 156 (8.917863, 47.73714) 8.917863 47.73714 2018-05-15 08:44:14.813975
#> 157 (8.917616, 47.73736) 8.917616 47.73736 2018-05-15 08:48:19.071676
#> 158 (8.917286, 47.73773) 8.917286 47.73773 2018-05-15 08:52:19.564013
#> 159  (8.917398, 47.7381) 8.917398 47.73810 2018-05-15 08:56:21.344632
#> 160  (8.918208, 47.7382) 8.918208 47.73820 2018-05-15 09:00:23.368380
#> 161 (8.918789, 47.73854) 8.918789 47.73854 2018-05-15 09:04:26.100668
#> 162 (8.919122, 47.73894) 8.919122 47.73894 2018-05-15 09:08:25.167637
#> 163  (8.91939, 47.73931) 8.919390 47.73931 2018-05-15 09:12:28.209587
#> 164 (8.919751, 47.73956) 8.919751 47.73956 2018-05-15 09:16:28.927083
#> 165 (8.920318, 47.73963) 8.920318 47.73963 2018-05-15 09:20:31.679605
#> 166 (8.921043, 47.73946) 8.921043 47.73946 2018-05-15 09:24:31.413633
#> 167 (8.922185, 47.73946) 8.922185 47.73946 2018-05-15 09:28:32.191066
#> 168  (8.923545, 47.7397) 8.923545 47.73970 2018-05-15 09:32:33.198207
#> 169  (8.924765, 47.7401) 8.924765 47.74010 2018-05-15 09:36:34.711327
#> 170 (8.925489, 47.74059) 8.925489 47.74059 2018-05-15 09:40:37.804982
#> 171  (8.925416, 47.7414) 8.925416 47.74140 2018-05-15 09:44:42.073742
#> 172 (8.926139, 47.74191) 8.926139 47.74191 2018-05-15 09:48:41.041020
#> 173 (8.926767, 47.74252) 8.926767 47.74252 2018-05-15 09:52:41.796959
#> 174  (8.927155, 47.7433) 8.927155 47.74330 2018-05-15 09:56:44.539561
#> 175  (8.92733, 47.74423) 8.927330 47.74423 2018-05-15 10:00:45.220897
#> 176 (8.927475, 47.74503) 8.927475 47.74503 2018-05-15 10:04:46.720574
#> 177  (8.927792, 47.7457) 8.927792 47.74570 2018-05-15 10:08:48.670071
#> 178 (8.928526, 47.74622) 8.928526 47.74622 2018-05-15 10:12:49.576386
#> 179  (8.929701, 47.7466) 8.929701 47.74660 2018-05-15 10:16:53.089646
#> 180 (8.931197, 47.74689) 8.931197 47.74689 2018-05-15 10:20:53.152303
#> 181  (8.93272, 47.74713) 8.932720 47.74713 2018-05-15 10:24:53.737716
#> 182 (8.933748, 47.74754) 8.933748 47.74754 2018-05-15 10:28:56.424901
#> 183  (8.934578, 47.7481) 8.934578 47.74810 2018-05-15 10:32:57.378423
#> 184  (8.935214, 47.7486) 8.935214 47.74860 2018-05-15 10:36:58.565655
#> 185 (8.935704, 47.74888) 8.935704 47.74888 2018-05-15 10:41:00.215390
#> 186  (8.936141, 47.7488) 8.936141 47.74880 2018-05-15 10:45:03.898446
#> 187 (8.936794, 47.74868) 8.936794 47.74868 2018-05-15 10:49:02.697466
#> 188  (8.937392, 47.7486) 8.937392 47.74860 2018-05-15 10:53:05.471171
#> 189 (8.938371, 47.74843) 8.938371 47.74843 2018-05-15 10:57:07.030526
#> 190 (8.939673, 47.74822) 8.939673 47.74822 2018-05-15 11:01:08.346776
#> 191 (8.940839, 47.74807) 8.940839 47.74807 2018-05-15 11:05:07.867436
#> 192   (8.941238, 47.748) 8.941238 47.74800 2018-05-15 11:09:11.990891
#> 193 (8.942187, 47.74793) 8.942187 47.74793 2018-05-15 11:13:12.597235
#> 194 (8.943496, 47.74777) 8.943496 47.74777 2018-05-15 11:17:15.062950
#> 195 (8.945177, 47.74748) 8.945177 47.74748 2018-05-15 11:21:16.489968
#> 196 (8.947286, 47.74701) 8.947286 47.74701 2018-05-15 11:25:15.560625
#> 197 (8.949463, 47.74647) 8.949463 47.74647 2018-05-15 11:29:16.994610
#> 198 (8.952061, 47.74614) 8.952061 47.74614 2018-05-15 11:33:20.979384
#> 199 (8.954235, 47.74536) 8.954235 47.74536 2018-05-15 11:37:22.449053
#> 200 (8.956099, 47.74432) 8.956099 47.74432 2018-05-15 11:41:22.591239
#> 201 (8.958176, 47.74347) 8.958176 47.74347 2018-05-15 11:45:25.921157
#> 202 (8.961297, 47.74283) 8.961297 47.74283 2018-05-15 11:49:24.756413
#> 203  (8.96423, 47.74211) 8.964230 47.74211 2018-05-15 11:53:27.183872
#> 204 (8.966908, 47.74162) 8.966908 47.74162 2018-05-15 11:57:28.636615
#> 205 (8.969312, 47.74126) 8.969312 47.74126 2018-05-15 12:01:29.841463
#> 206 (8.971508, 47.74091) 8.971508 47.74091 2018-05-15 12:05:30.304964
#> 207  (8.97383, 47.74081) 8.973830 47.74081 2018-05-15 12:09:34.541600
#> 208  (8.97486, 47.74158) 8.974860 47.74158 2018-05-15 12:13:33.367778
#> 209  (8.97479, 47.74306) 8.974790 47.74306 2018-05-15 12:17:36.392157
#> 210 (8.974003, 47.74461) 8.974003 47.74461 2018-05-15 12:21:36.699414
#> 211 (8.972736, 47.74569) 8.972736 47.74569 2018-05-15 12:25:40.533443
#> 212 (8.970559, 47.74633) 8.970559 47.74633 2018-05-15 12:29:40.033745
#> 213 (8.969987, 47.74732) 8.969987 47.74732 2018-05-15 12:33:42.798597
#> 214 (8.970193, 47.74822) 8.970193 47.74822 2018-05-15 12:37:44.270246
#> 215 (8.971177, 47.74917) 8.971177 47.74917 2018-05-15 12:41:46.328587
#> 216 (8.972788, 47.75038) 8.972788 47.75038 2018-05-15 12:45:48.052170
#> 217 (8.974457, 47.75194) 8.974457 47.75194 2018-05-15 12:49:48.709585
#> 218 (8.975088, 47.75318) 8.975088 47.75318 2018-05-15 12:53:48.941821
#> 219 (8.974998, 47.75419) 8.974998 47.75419 2018-05-15 12:57:51.736988
#> 220 (8.974159, 47.75525) 8.974159 47.75525 2018-05-15 13:01:52.038555
#> 221 (8.972715, 47.75628) 8.972715 47.75628 2018-05-15 13:05:53.250144
#> 222 (8.971046, 47.75668) 8.971046 47.75668 2018-05-15 13:09:55.640757
#> 223 (8.969162, 47.75745) 8.969162 47.75745 2018-05-15 13:13:57.480460
#> 224 (8.967868, 47.75848) 8.967868 47.75848 2018-05-15 13:18:00.596955
#> 225 (8.966852, 47.75982) 8.966852 47.75982 2018-05-15 13:22:00.971234
#> 226 (8.965497, 47.76137) 8.965497 47.76137 2018-05-15 13:26:03.515531
#> 227 (8.962945, 47.76287) 8.962945 47.76287 2018-05-15 13:30:04.069746
#> 228  (8.959607, 47.7636) 8.959607 47.76360 2018-05-15 13:34:07.006698
#> 229 (8.956101, 47.76385) 8.956101 47.76385 2018-05-15 13:38:04.828377
#> 230 (8.952892, 47.76394) 8.952892 47.76394 2018-05-15 13:42:08.672040
#> 231  (8.94999, 47.76425) 8.949990 47.76425 2018-05-15 13:46:10.456284
#> 232  (8.947114, 47.7651) 8.947114 47.76510 2018-05-15 13:50:11.571585
#> 233 (8.944792, 47.76622) 8.944792 47.76622 2018-05-15 13:54:11.359732
#> 234 (8.941945, 47.76612) 8.941945 47.76612 2018-05-15 13:58:15.789150
#> 235 (8.939092, 47.76581) 8.939092 47.76581 2018-05-15 14:02:15.946422
#> 236 (8.936428, 47.76566) 8.936428 47.76566 2018-05-15 14:06:16.389307
#> 237 (8.933972, 47.76573) 8.933972 47.76573 2018-05-15 14:10:17.192129
#> 238  (8.93165, 47.76602) 8.931650 47.76602 2018-05-15 14:14:21.618717
#> 239 (8.930177, 47.76607) 8.930177 47.76607 2018-05-15 14:18:19.040766
#> 240 (8.929043, 47.76602) 8.929043 47.76602 2018-05-15 14:22:24.196560
#> 241 (8.928231, 47.76588) 8.928231 47.76588 2018-05-15 14:26:23.637514
#> 242 (8.927796, 47.76572) 8.927796 47.76572 2018-05-15 14:30:27.423720
#> 243 (8.927361, 47.76582) 8.927361 47.76582 2018-05-15 14:34:28.727044
#> 244  (8.927429, 47.7663) 8.927429 47.76630 2018-05-15 14:38:29.340269
#> 245 (8.927211, 47.76668) 8.927211 47.76668 2018-05-15 14:42:29.756380
#> 246  (8.92647, 47.76692) 8.926470 47.76692 2018-05-15 14:46:30.818439
#> 247 (8.925318, 47.76698) 8.925318 47.76698 2018-05-15 14:50:34.996217
#> 248 (8.924665, 47.76654) 8.924665 47.76654 2018-05-15 14:54:35.981282
#> 249 (8.923425, 47.76654) 8.923425 47.76654 2018-05-15 14:58:35.102100
#> 250   (8.9225, 47.76624) 8.922500 47.76624 2018-05-15 15:02:38.372865
#> 251 (8.922315, 47.76583) 8.922315 47.76583 2018-05-15 15:06:38.172376
#> 252 (8.922796, 47.76546) 8.922796 47.76546 2018-05-15 15:10:42.387372
#> 253 (8.922796, 47.76473) 8.922796 47.76473 2018-05-15 15:14:41.628728
#> 254 (8.923268, 47.76418) 8.923268 47.76418 2018-05-15 15:18:44.694897
#> 255 (8.923676, 47.76351) 8.923676 47.76351 2018-05-15 15:22:44.907991
#> 256 (8.924168, 47.76313) 8.924168 47.76313 2018-05-15 15:26:48.117283
#> 257 (8.924748, 47.76324) 8.924748 47.76324 2018-05-15 15:30:49.783762
#> 258 (8.925038, 47.76363) 8.925038 47.76363 2018-05-15 15:34:51.333863
#> 259 (8.925119, 47.76281) 8.925119 47.76281 2018-05-15 15:38:52.628372
#> 260  (8.924774, 47.7621) 8.924774 47.76210 2018-05-15 15:42:54.265388
#> 261 (8.924197, 47.76165) 8.924197 47.76165 2018-05-15 15:46:54.760231
#> 262 (8.923442, 47.76133) 8.923442 47.76133 2018-05-15 15:50:55.186109
#> 263 (8.922354, 47.76075) 8.922354 47.76075 2018-05-15 15:54:59.507446
#> 264 (8.920905, 47.75982) 8.920905 47.75982 2018-05-15 15:59:00.091301
#> 265  (8.919291, 47.7586) 8.919291 47.75860 2018-05-15 16:03:00.287453
#> 266  (8.91799, 47.75758) 8.917990 47.75758 2018-05-15 16:07:01.991550
#> 267  (8.917182, 47.7569) 8.917182 47.75690 2018-05-15 16:11:04.181335
#> 268 (8.916753, 47.75636) 8.916753 47.75636 2018-05-15 16:15:04.908397
#> 269 (8.916608, 47.75573) 8.916608 47.75573 2018-05-15 16:19:06.238394
#> 270  (8.91618, 47.75574) 8.916180 47.75574 2018-05-15 16:23:09.845695
#> 271 (8.916141, 47.75607) 8.916141 47.75607 2018-05-15 16:27:11.525078
#> 272 (8.916439, 47.75629) 8.916439 47.75629 2018-05-15 16:31:11.221723
#> 273 (8.916576, 47.75601) 8.916576 47.75601 2018-05-15 16:35:12.367775
#> 274 (8.916576, 47.75601) 8.916576 47.75601 2018-05-15 16:39:14.268187
#> 275 (8.916871, 47.75578) 8.916871 47.75578 2018-05-15 16:43:17.747497
#> 276 (8.917027, 47.75562) 8.917027 47.75562 2018-05-15 16:47:16.869316
#> 277 (8.916575, 47.75573) 8.916575 47.75573 2018-05-15 16:51:20.377056
#> 278 (8.915371, 47.75612) 8.915371 47.75612 2018-05-15 16:55:20.234575
#> 279 (8.916024, 47.75598) 8.916024 47.75598 2018-05-15 16:59:23.116363
#> 280 (8.916799, 47.75554) 8.916799 47.75554 2018-05-15 17:03:24.200968
#> 281   (8.917335, 47.755) 8.917335 47.75500 2018-05-15 17:07:26.734737
#> 282  (8.918142, 47.7545) 8.918142 47.75450 2018-05-15 17:11:26.756195
#> 283  (8.919704, 47.7541) 8.919704 47.75410 2018-05-15 17:15:29.154650
#> 284  (8.92101, 47.75327) 8.921010 47.75327 2018-05-15 17:19:27.978102
#> 285 (8.923396, 47.75295) 8.923396 47.75295 2018-05-15 17:23:32.006437
#> 286 (8.925568, 47.75296) 8.925568 47.75296 2018-05-15 17:27:33.351746
#> 287 (8.927329, 47.75305) 8.927329 47.75305 2018-05-15 17:31:34.821926
#> 288   (8.92864, 47.7529) 8.928640 47.75290 2018-05-15 17:35:35.248992
#> 289 (8.929148, 47.75192) 8.929148 47.75192 2018-05-15 17:39:36.728973
#> 290 (8.928078, 47.75002) 8.928078 47.75002 2018-05-15 17:43:37.002882
#> 291  (8.92617, 47.74821) 8.926170 47.74821 2018-05-15 17:47:39.903844
#> 292 (8.925114, 47.74667) 8.925114 47.74667 2018-05-15 17:51:42.158933
#> 293 (8.925525, 47.74535) 8.925525 47.74535 2018-05-15 17:55:42.111845
#> 294 (8.925162, 47.74398) 8.925162 47.74398 2018-05-15 17:59:46.036300
#> 295 (8.923983, 47.74293) 8.923983 47.74293 2018-05-15 18:03:48.724115
#> 296 (8.923355, 47.74174) 8.923355 47.74174 2018-05-15 18:07:46.580075
#> 297 (8.922828, 47.74056) 8.922828 47.74056 2018-05-15 18:11:48.546778
#> 298 (8.922224, 47.73955) 8.922224 47.73955 2018-05-15 18:15:50.898211
#> 299 (8.922079, 47.73877) 8.922079 47.73877 2018-05-15 18:19:52.724342
#> 300 (8.922567, 47.73798) 8.922567 47.73798 2018-05-15 18:23:54.255801
#> 301 (8.923233, 47.73742) 8.923233 47.73742 2018-05-15 18:27:55.455560
#> 302  (8.92308, 47.73679) 8.923080 47.73679 2018-05-15 18:31:57.506520
#> 303 (8.922009, 47.73609) 8.922009 47.73609 2018-05-15 18:35:58.637337
#> 304 (8.920993, 47.73565) 8.920993 47.73565 2018-05-15 18:40:00.045075
#> 305 (8.939787, 47.72868) 8.939787 47.72868 2018-05-15 06:59:59.636287
#> 306 (8.939642, 47.72907) 8.939642 47.72907 2018-05-15 07:04:00.117243
#> 307 (8.939853, 47.72959) 8.939853 47.72959 2018-05-15 07:08:04.750756
#> 308 (8.940105, 47.73005) 8.940105 47.73005 2018-05-15 07:12:02.634175
#> 309 (8.940238, 47.73042) 8.940238 47.73042 2018-05-15 07:16:05.389819
#> 310 (8.940246, 47.73081) 8.940246 47.73081 2018-05-15 07:20:05.369172
#> 311  (8.940572, 47.7312) 8.940572 47.73120 2018-05-15 07:24:06.913951
#> 312   (8.9407, 47.73144) 8.940700 47.73144 2018-05-15 07:28:06.985471
#> 313 (8.940792, 47.73168) 8.940792 47.73168 2018-05-15 07:32:09.806045
#> 314   (8.9409, 47.73193) 8.940900 47.73193 2018-05-15 07:36:10.314162
#> 315 (8.941011, 47.73206) 8.941011 47.73206 2018-05-15 07:40:10.961310
#> 316 (8.941156, 47.73235) 8.941156 47.73235 2018-05-15 07:44:13.795185
#> 317 (8.941417, 47.73248) 8.941417 47.73248 2018-05-15 07:48:14.249375
#> 318  (8.941707, 47.7324) 8.941707 47.73240 2018-05-15 07:52:17.074906
#> 319  (8.941951, 47.7322) 8.941951 47.73220 2018-05-15 07:56:17.117854
#> 320 (8.942117, 47.73205) 8.942117 47.73205 2018-05-15 08:00:18.042339
#> 321 (8.942552, 47.73209) 8.942552 47.73209 2018-05-15 08:04:19.458911
#> 322 (8.943017, 47.73218) 8.943017 47.73218 2018-05-15 08:08:18.202832
#> 323 (8.943212, 47.73235) 8.943212 47.73235 2018-05-15 08:12:20.655024
#> 324 (8.943262, 47.73253) 8.943262 47.73253 2018-05-15 08:16:20.369044
#> 325 (8.943431, 47.73262) 8.943431 47.73262 2018-05-15 08:20:25.054933
#> 326  (8.94383, 47.73267) 8.943830 47.73267 2018-05-15 08:24:24.511894
#> 327 (8.943981, 47.73253) 8.943981 47.73253 2018-05-15 08:28:23.292874
#> 328   (8.94409, 47.7323) 8.944090 47.73230 2018-05-15 08:32:26.498149
#> 329 (8.944039, 47.73207) 8.944039 47.73207 2018-05-15 08:36:28.157185
#> 330 (8.943722, 47.73192) 8.943722 47.73192 2018-05-15 08:40:30.203019
#> 331 (8.943468, 47.73192) 8.943468 47.73192 2018-05-15 08:44:31.526079
#> 332  (8.943247, 47.7319) 8.943247 47.73190 2018-05-15 08:48:32.243129
#> 333 (8.942937, 47.73195) 8.942937 47.73195 2018-05-15 08:52:33.092540
#> 334 (8.942674, 47.73205) 8.942674 47.73205 2018-05-15 08:56:34.912570
#> 335 (8.942581, 47.73216) 8.942581 47.73216 2018-05-15 09:00:35.348577
#> 336 (8.942581, 47.73233) 8.942581 47.73233 2018-05-15 09:04:36.482333
#> 337 (8.942516, 47.73257) 8.942516 47.73257 2018-05-15 09:08:37.128171
#> 338 (8.942348, 47.73263) 8.942348 47.73263 2018-05-15 09:12:40.509000
#> 339 (8.942296, 47.73268) 8.942296 47.73268 2018-05-15 09:16:40.183602
#> 340  (8.942495, 47.7328) 8.942495 47.73280 2018-05-15 09:20:40.462875
#> 341 (8.942785, 47.73283) 8.942785 47.73283 2018-05-15 09:24:43.044000
#> 342 (8.943344, 47.73283) 8.943344 47.73283 2018-05-15 09:28:43.664164
#> 343 (8.943825, 47.73285) 8.943825 47.73285 2018-05-15 09:32:43.436789
#> 344   (8.944259, 47.733) 8.944259 47.73300 2018-05-15 09:36:45.755334
#> 345 (8.944623, 47.73332) 8.944623 47.73332 2018-05-15 09:40:47.952697
#> 346 (8.944696, 47.73366) 8.944696 47.73366 2018-05-15 09:44:47.906567
#> 347  (8.945423, 47.7337) 8.945423 47.73370 2018-05-15 09:48:49.830337
#> 348  (8.94621, 47.73385) 8.946210 47.73385 2018-05-15 09:52:50.795372
#> 349 (8.946671, 47.73401) 8.946671 47.73401 2018-05-15 09:56:50.189469
#> 350 (8.946831, 47.73417) 8.946831 47.73417 2018-05-15 10:00:53.230359
#> 351 (8.947302, 47.73446) 8.947302 47.73446 2018-05-15 10:04:55.092491
#> 352 (8.947985, 47.73466) 8.947985 47.73466 2018-05-15 10:08:54.138597
#> 353 (8.948885, 47.73467) 8.948885 47.73467 2018-05-15 10:12:56.442235
#> 354 (8.949722, 47.73464) 8.949722 47.73464 2018-05-15 10:16:58.941830
#> 355 (8.950481, 47.73462) 8.950481 47.73462 2018-05-15 10:20:57.947191
#> 356  (8.95146, 47.73462) 8.951460 47.73462 2018-05-15 10:25:01.264811
#> 357 (8.952329, 47.73455) 8.952329 47.73455 2018-05-15 10:29:00.375951
#> 358 (8.953391, 47.73433) 8.953391 47.73433 2018-05-15 10:33:02.577331
#> 359 (8.954254, 47.73396) 8.954254 47.73396 2018-05-15 10:37:04.576173
#> 360 (8.954821, 47.73354) 8.954821 47.73354 2018-05-15 10:41:04.618848
#> 361 (8.955263, 47.73326) 8.955263 47.73326 2018-05-15 10:45:06.340006
#> 362 (8.955988, 47.73328) 8.955988 47.73328 2018-05-15 10:49:06.794555
#> 363  (8.956458, 47.7334) 8.956458 47.73340 2018-05-15 10:53:08.520195
#> 364 (8.956818, 47.73369) 8.956818 47.73369 2018-05-15 10:57:08.812954
#> 365 (8.957229, 47.73398) 8.957229 47.73398 2018-05-15 11:01:10.698200
#> 366 (8.957841, 47.73399) 8.957841 47.73399 2018-05-15 11:05:11.981431
#> 367 (8.958384, 47.73414) 8.958384 47.73414 2018-05-15 11:09:13.996612
#> 368 (8.958432, 47.73448) 8.958432 47.73448 2018-05-15 11:13:13.629700
#> 369 (8.958787, 47.73473) 8.958787 47.73473 2018-05-15 11:17:14.906635
#> 370 (8.959383, 47.73478) 8.959383 47.73478 2018-05-15 11:21:16.476073
#> 371 (8.959821, 47.73463) 8.959821 47.73463 2018-05-15 11:25:20.221409
#> 372 (8.960763, 47.73439) 8.960763 47.73439 2018-05-15 11:29:17.717775
#> 373  (8.961585, 47.7341) 8.961585 47.73410 2018-05-15 11:33:21.184670
#> 374 (8.962404, 47.73403) 8.962404 47.73403 2018-05-15 11:37:20.860746
#> 375  (8.963175, 47.7341) 8.963175 47.73410 2018-05-15 11:41:22.558886
#> 376 (8.963841, 47.73417) 8.963841 47.73417 2018-05-15 11:45:20.992771
#> 377 (8.964638, 47.73424) 8.964638 47.73424 2018-05-15 11:49:25.669662
#> 378  (8.965652, 47.7343) 8.965652 47.73430 2018-05-15 11:53:24.779064
#> 379  (8.96646, 47.73417) 8.966460 47.73417 2018-05-15 11:57:26.856449
#> 380  (8.967046, 47.7339) 8.967046 47.73390 2018-05-15 12:01:27.750536
#> 381 (8.967539, 47.73358) 8.967539 47.73358 2018-05-15 12:05:30.264261
#> 382 (8.968191, 47.73329) 8.968191 47.73329 2018-05-15 12:09:30.366362
#> 383 (8.968907, 47.73338) 8.968907 47.73338 2018-05-15 12:13:30.334724
#> 384 (8.969539, 47.73368) 8.969539 47.73368 2018-05-15 12:17:31.780115
#> 385   (8.970077, 47.734) 8.970077 47.73400 2018-05-15 12:21:34.234659
#> 386 (8.970511, 47.73421) 8.970511 47.73421 2018-05-15 12:25:36.971302
#> 387  (8.970728, 47.7345) 8.970728 47.73450 2018-05-15 12:29:36.739850
#> 388 (8.970651, 47.73459) 8.970651 47.73459 2018-05-15 12:33:37.565245
#> 389 (8.970686, 47.73451) 8.970686 47.73451 2018-05-15 12:37:38.677310
#> 390 (8.970644, 47.73445) 8.970644 47.73445 2018-05-15 12:41:40.315721
#> 391  (8.970496, 47.7345) 8.970496 47.73450 2018-05-15 12:45:42.878278
#> 392 (8.970858, 47.73436) 8.970858 47.73436 2018-05-15 12:49:42.889331
#> 393 (8.971187, 47.73437) 8.971187 47.73437 2018-05-15 12:53:44.362516
#> 394 (8.971612, 47.73426) 8.971612 47.73426 2018-05-15 12:57:44.489384
#> 395 (8.972194, 47.73423) 8.972194 47.73423 2018-05-15 13:01:46.514625
#> 396 (8.972954, 47.73436) 8.972954 47.73436 2018-05-15 13:05:45.172058
#> 397  (8.97386, 47.73443) 8.973860 47.73443 2018-05-15 13:09:49.860785
#> 398 (8.974835, 47.73441) 8.974835 47.73441 2018-05-15 13:13:48.556840
#> 399 (8.975659, 47.73451) 8.975659 47.73451 2018-05-15 13:17:50.521399
#> 400 (8.976084, 47.73474) 8.976084 47.73474 2018-05-15 13:21:50.362326
#> 401 (8.976168, 47.73497) 8.976168 47.73497 2018-05-15 13:25:51.166356
#> 402 (8.976422, 47.73492) 8.976422 47.73492 2018-05-15 13:29:54.489474
#> 403 (8.976216, 47.73514) 8.976216 47.73514 2018-05-15 13:33:55.285872
#> 404 (8.976635, 47.73503) 8.976635 47.73503 2018-05-15 13:37:55.794087
#> 405 (8.976898, 47.73499) 8.976898 47.73499 2018-05-15 13:41:56.032696
#> 406 (8.976771, 47.73499) 8.976771 47.73499 2018-05-15 13:45:58.882898
#> 407 (8.976698, 47.73453) 8.976698 47.73453 2018-05-15 13:50:00.624386
#> 408 (8.976698, 47.73399) 8.976698 47.73399 2018-05-15 13:54:01.662190
#> 409 (8.976397, 47.73346) 8.976397 47.73346 2018-05-15 13:58:02.765029
#> 410 (8.976388, 47.73298) 8.976388 47.73298 2018-05-15 14:02:05.187101
#> 411 (8.976566, 47.73252) 8.976566 47.73252 2018-05-15 14:06:04.021298
#> 412 (8.976241, 47.73193) 8.976241 47.73193 2018-05-15 14:10:06.390619
#> 413 (8.975988, 47.73125) 8.975988 47.73125 2018-05-15 14:14:08.113441
#> 414 (8.975225, 47.73093) 8.975225 47.73093 2018-05-15 14:18:07.662658
#> 415 (8.974689, 47.73035) 8.974689 47.73035 2018-05-15 14:22:09.862999
#> 416 (8.974302, 47.72959) 8.974302 47.72959 2018-05-15 14:26:08.697848
#> 417  (8.973427, 47.7292) 8.973427 47.72920 2018-05-15 14:30:11.333275
#> 418 (8.971036, 47.72866) 8.971036 47.72866 2018-05-15 14:34:13.670732
#> 419 (8.969286, 47.72844) 8.969286 47.72844 2018-05-15 14:38:13.834596
#> 420 (8.968098, 47.72821) 8.968098 47.72821 2018-05-15 14:42:13.512051
#> 421 (8.967068, 47.72804) 8.967068 47.72804 2018-05-15 14:46:15.995625
#> 422 (8.965728, 47.72817) 8.965728 47.72817 2018-05-15 14:50:18.450314
#> 423 (8.964858, 47.72871) 8.964858 47.72871 2018-05-15 14:54:17.707817
#> 424 (8.963929, 47.72876) 8.963929 47.72876 2018-05-15 14:58:18.775724
#> 425 (8.962987, 47.72885) 8.962987 47.72885 2018-05-15 15:02:19.839735
#> 426 (8.961797, 47.72891) 8.961797 47.72891 2018-05-15 15:06:21.623984
#> 427 (8.960173, 47.72877) 8.960173 47.72877 2018-05-15 15:10:23.630927
#> 428 (8.958941, 47.72911) 8.958941 47.72911 2018-05-15 15:14:24.496638
#> 429 (8.957647, 47.72875) 8.957647 47.72875 2018-05-15 15:18:26.711852
#> 430 (8.956387, 47.72856) 8.956387 47.72856 2018-05-15 15:22:25.406311
#> 431 (8.955102, 47.72877) 8.955102 47.72877 2018-05-15 15:26:28.741696
#> 432 (8.953734, 47.72927) 8.953734 47.72927 2018-05-15 15:30:29.361044
#> 433  (8.95243, 47.72912) 8.952430 47.72912 2018-05-15 15:34:31.810392
#> 434 (8.951253, 47.72867) 8.951253 47.72867 2018-05-15 15:38:33.924942
#> 435 (8.950139, 47.72874) 8.950139 47.72874 2018-05-15 15:42:33.707971
#> 436  (8.948968, 47.7289) 8.948968 47.72890 2018-05-15 15:46:35.051258
#> 437 (8.947727, 47.72873) 8.947727 47.72873 2018-05-15 15:50:37.250378
#> 438 (8.946857, 47.72868) 8.946857 47.72868 2018-05-15 15:54:37.047001
#> 439 (8.946249, 47.72916) 8.946249 47.72916 2018-05-15 15:58:38.174949
#> 440  (8.945336, 47.7295) 8.945336 47.72950 2018-05-15 16:02:38.235137
#> 441 (8.944429, 47.72967) 8.944429 47.72967 2018-05-15 16:06:41.979844
#> 442 (8.943817, 47.72969) 8.943817 47.72969 2018-05-15 16:10:40.517418
#> 443   (8.9434, 47.72961) 8.943400 47.72961 2018-05-15 16:14:41.757091
#> 444 (8.943135, 47.72955) 8.943135 47.72955 2018-05-15 16:18:45.237168
#> 445 (8.942844, 47.72946) 8.942844 47.72946 2018-05-15 16:22:45.349187
#> 446 (8.942441, 47.72938) 8.942441 47.72938 2018-05-15 16:26:45.259415
#> 447 (8.941975, 47.72935) 8.941975 47.72935 2018-05-15 16:30:47.501894
#> 448 (8.941758, 47.72939) 8.941758 47.72939 2018-05-15 16:34:47.579344
#> 449 (8.941592, 47.72958) 8.941592 47.72958 2018-05-15 16:38:50.438985
#> 450 (8.941301, 47.72962) 8.941301 47.72962 2018-05-15 16:42:49.419842
#> 451 (8.941074, 47.72949) 8.941074 47.72949 2018-05-15 16:46:51.251741
#> 452 (8.940984, 47.72925) 8.940984 47.72925 2018-05-15 16:50:54.129838
#> 453 (8.940893, 47.72894) 8.940893 47.72894 2018-05-15 16:54:53.861109
#> 454 (8.940732, 47.72878) 8.940732 47.72878 2018-05-15 16:58:56.490841
#> 455 (8.940556, 47.72868) 8.940556 47.72868 2018-05-15 17:02:56.354562
#> 456 (8.940313, 47.72864) 8.940313 47.72864 2018-05-15 17:06:56.686083
#> 457 (8.940044, 47.72863) 8.940044 47.72863 2018-05-15 17:11:00.026924
#> 458   (8.9399, 47.72865) 8.939900 47.72865 2018-05-15 17:15:00.973202