Get frame times from frames

This function extracts the timestamps associated with each frame from a moveVis object created using frames_spatial or frames_graph and returns them as a vector.

Usage

get_frametimes(frames)

Arguments

frames

an object of class moveVis created with frames_spatial.

Value

A POSIXct vector of timestamps representing the time associated with each frame.

See also

Examples

library(moveVis)
library(move2)
library(terra)

data("move_data", package = "moveVis")
r <- readRDS(example_data(file = "raster_NDVI.rds"))

# align movement
m <- align_move(move_data, res = units::set_units(4, "min"))
#> Temporal resolution of 4 [min] is used to align trajectories.

# create frames 
frames <- frames_spatial(m, r, r_type = "gradient", fade_raster = TRUE) 
#> Processing input data...
#> Approximated animation duration: ≈ 7.52s at 25 fps using 188 frames
#> CRS (geodetic):   WGS 84
#> Assigning raster maps to frames...

# get frametimes
get_frametimes(frames)
#>   [1] "2018-05-15 07:03:59 UTC" "2018-05-15 07:07:59 UTC"
#>   [3] "2018-05-15 07:11:59 UTC" "2018-05-15 07:15:59 UTC"
#>   [5] "2018-05-15 07:19:59 UTC" "2018-05-15 07:23:59 UTC"
#>   [7] "2018-05-15 07:27:59 UTC" "2018-05-15 07:31:59 UTC"
#>   [9] "2018-05-15 07:35:59 UTC" "2018-05-15 07:39:59 UTC"
#>  [11] "2018-05-15 07:43:59 UTC" "2018-05-15 07:47:59 UTC"
#>  [13] "2018-05-15 07:51:59 UTC" "2018-05-15 07:55:59 UTC"
#>  [15] "2018-05-15 07:59:59 UTC" "2018-05-15 08:03:59 UTC"
#>  [17] "2018-05-15 08:07:59 UTC" "2018-05-15 08:11:59 UTC"
#>  [19] "2018-05-15 08:15:59 UTC" "2018-05-15 08:19:59 UTC"
#>  [21] "2018-05-15 08:23:59 UTC" "2018-05-15 08:27:59 UTC"
#>  [23] "2018-05-15 08:31:59 UTC" "2018-05-15 08:35:59 UTC"
#>  [25] "2018-05-15 08:39:59 UTC" "2018-05-15 08:43:59 UTC"
#>  [27] "2018-05-15 08:47:59 UTC" "2018-05-15 08:51:59 UTC"
#>  [29] "2018-05-15 08:55:59 UTC" "2018-05-15 08:59:59 UTC"
#>  [31] "2018-05-15 09:03:59 UTC" "2018-05-15 09:07:59 UTC"
#>  [33] "2018-05-15 09:11:59 UTC" "2018-05-15 09:15:59 UTC"
#>  [35] "2018-05-15 09:19:59 UTC" "2018-05-15 09:23:59 UTC"
#>  [37] "2018-05-15 09:27:59 UTC" "2018-05-15 09:31:59 UTC"
#>  [39] "2018-05-15 09:35:59 UTC" "2018-05-15 09:39:59 UTC"
#>  [41] "2018-05-15 09:43:59 UTC" "2018-05-15 09:47:59 UTC"
#>  [43] "2018-05-15 09:51:59 UTC" "2018-05-15 09:55:59 UTC"
#>  [45] "2018-05-15 09:59:59 UTC" "2018-05-15 10:03:59 UTC"
#>  [47] "2018-05-15 10:07:59 UTC" "2018-05-15 10:11:59 UTC"
#>  [49] "2018-05-15 10:15:59 UTC" "2018-05-15 10:19:59 UTC"
#>  [51] "2018-05-15 10:23:59 UTC" "2018-05-15 10:27:59 UTC"
#>  [53] "2018-05-15 10:31:59 UTC" "2018-05-15 10:35:59 UTC"
#>  [55] "2018-05-15 10:39:59 UTC" "2018-05-15 10:43:59 UTC"
#>  [57] "2018-05-15 10:47:59 UTC" "2018-05-15 10:51:59 UTC"
#>  [59] "2018-05-15 10:55:59 UTC" "2018-05-15 10:59:59 UTC"
#>  [61] "2018-05-15 11:03:59 UTC" "2018-05-15 11:07:59 UTC"
#>  [63] "2018-05-15 11:11:59 UTC" "2018-05-15 11:15:59 UTC"
#>  [65] "2018-05-15 11:19:59 UTC" "2018-05-15 11:23:59 UTC"
#>  [67] "2018-05-15 11:27:59 UTC" "2018-05-15 11:31:59 UTC"
#>  [69] "2018-05-15 11:35:59 UTC" "2018-05-15 11:39:59 UTC"
#>  [71] "2018-05-15 11:43:59 UTC" "2018-05-15 11:47:59 UTC"
#>  [73] "2018-05-15 11:51:59 UTC" "2018-05-15 11:55:59 UTC"
#>  [75] "2018-05-15 11:59:59 UTC" "2018-05-15 12:03:59 UTC"
#>  [77] "2018-05-15 12:07:59 UTC" "2018-05-15 12:11:59 UTC"
#>  [79] "2018-05-15 12:15:59 UTC" "2018-05-15 12:19:59 UTC"
#>  [81] "2018-05-15 12:23:59 UTC" "2018-05-15 12:27:59 UTC"
#>  [83] "2018-05-15 12:31:59 UTC" "2018-05-15 12:35:59 UTC"
#>  [85] "2018-05-15 12:39:59 UTC" "2018-05-15 12:43:59 UTC"
#>  [87] "2018-05-15 12:47:59 UTC" "2018-05-15 12:51:59 UTC"
#>  [89] "2018-05-15 12:55:59 UTC" "2018-05-15 12:59:59 UTC"
#>  [91] "2018-05-15 13:03:59 UTC" "2018-05-15 13:07:59 UTC"
#>  [93] "2018-05-15 13:11:59 UTC" "2018-05-15 13:15:59 UTC"
#>  [95] "2018-05-15 13:19:59 UTC" "2018-05-15 13:23:59 UTC"
#>  [97] "2018-05-15 13:27:59 UTC" "2018-05-15 13:31:59 UTC"
#>  [99] "2018-05-15 13:35:59 UTC" "2018-05-15 13:39:59 UTC"
#> [101] "2018-05-15 13:43:59 UTC" "2018-05-15 13:47:59 UTC"
#> [103] "2018-05-15 13:51:59 UTC" "2018-05-15 13:55:59 UTC"
#> [105] "2018-05-15 13:59:59 UTC" "2018-05-15 14:03:59 UTC"
#> [107] "2018-05-15 14:07:59 UTC" "2018-05-15 14:11:59 UTC"
#> [109] "2018-05-15 14:15:59 UTC" "2018-05-15 14:19:59 UTC"
#> [111] "2018-05-15 14:23:59 UTC" "2018-05-15 14:27:59 UTC"
#> [113] "2018-05-15 14:31:59 UTC" "2018-05-15 14:35:59 UTC"
#> [115] "2018-05-15 14:39:59 UTC" "2018-05-15 14:43:59 UTC"
#> [117] "2018-05-15 14:47:59 UTC" "2018-05-15 14:51:59 UTC"
#> [119] "2018-05-15 14:55:59 UTC" "2018-05-15 14:59:59 UTC"
#> [121] "2018-05-15 15:03:59 UTC" "2018-05-15 15:07:59 UTC"
#> [123] "2018-05-15 15:11:59 UTC" "2018-05-15 15:15:59 UTC"
#> [125] "2018-05-15 15:19:59 UTC" "2018-05-15 15:23:59 UTC"
#> [127] "2018-05-15 15:27:59 UTC" "2018-05-15 15:31:59 UTC"
#> [129] "2018-05-15 15:35:59 UTC" "2018-05-15 15:39:59 UTC"
#> [131] "2018-05-15 15:43:59 UTC" "2018-05-15 15:47:59 UTC"
#> [133] "2018-05-15 15:51:59 UTC" "2018-05-15 15:55:59 UTC"
#> [135] "2018-05-15 15:59:59 UTC" "2018-05-15 16:03:59 UTC"
#> [137] "2018-05-15 16:07:59 UTC" "2018-05-15 16:11:59 UTC"
#> [139] "2018-05-15 16:15:59 UTC" "2018-05-15 16:19:59 UTC"
#> [141] "2018-05-15 16:23:59 UTC" "2018-05-15 16:27:59 UTC"
#> [143] "2018-05-15 16:31:59 UTC" "2018-05-15 16:35:59 UTC"
#> [145] "2018-05-15 16:39:59 UTC" "2018-05-15 16:43:59 UTC"
#> [147] "2018-05-15 16:47:59 UTC" "2018-05-15 16:51:59 UTC"
#> [149] "2018-05-15 16:55:59 UTC" "2018-05-15 16:59:59 UTC"
#> [151] "2018-05-15 17:03:59 UTC" "2018-05-15 17:07:59 UTC"
#> [153] "2018-05-15 17:11:59 UTC" "2018-05-15 17:15:59 UTC"
#> [155] "2018-05-15 17:19:59 UTC" "2018-05-15 17:23:59 UTC"
#> [157] "2018-05-15 17:27:59 UTC" "2018-05-15 17:31:59 UTC"
#> [159] "2018-05-15 17:35:59 UTC" "2018-05-15 17:39:59 UTC"
#> [161] "2018-05-15 17:43:59 UTC" "2018-05-15 17:47:59 UTC"
#> [163] "2018-05-15 17:51:59 UTC" "2018-05-15 17:55:59 UTC"
#> [165] "2018-05-15 17:59:59 UTC" "2018-05-15 18:03:59 UTC"
#> [167] "2018-05-15 18:07:59 UTC" "2018-05-15 18:11:59 UTC"
#> [169] "2018-05-15 18:15:59 UTC" "2018-05-15 18:19:59 UTC"
#> [171] "2018-05-15 18:23:59 UTC" "2018-05-15 18:27:59 UTC"
#> [173] "2018-05-15 18:31:59 UTC" "2018-05-15 18:35:59 UTC"
#> [175] "2018-05-15 18:39:59 UTC" "2018-05-15 18:43:59 UTC"
#> [177] "2018-05-15 18:47:59 UTC" "2018-05-15 18:51:59 UTC"
#> [179] "2018-05-15 18:55:59 UTC" "2018-05-15 18:59:59 UTC"
#> [181] "2018-05-15 19:03:59 UTC" "2018-05-15 19:07:59 UTC"
#> [183] "2018-05-15 19:11:59 UTC" "2018-05-15 19:15:59 UTC"
#> [185] "2018-05-15 19:19:59 UTC" "2018-05-15 19:23:59 UTC"
#> [187] "2018-05-15 19:27:59 UTC" "2018-05-15 19:31:59 UTC"