Detect extern system libraries needed by moveVis

get_libraries trys to detect the libraries on your system that are needed by moveVis to be able to deal with different output file formats. get_libraries() searches for'convert' form the ImageMagick library (needed for .gif support), 'ffmpeg' from the FFmpeg library or 'avconv' from the 'libav-tools' library (both needed for video support). You can execute get_libraries() to make sure, these libraries are correctly installed on your system. It is recommended to have ImageMagick including 'convert' and FFmpeg including 'ffmpeg' installed to gain support of all available output file formats. The function's return can serve as conv_dir input to the animate_move() function.

get_imconvert is an deprecated alias function of get_libraries inlcuded for compatibility reasons that does the same as get_libraries, but is only checking for the convert tool of ImageMagick. It is recommended to use get_libraries instead.

get_libraries(lib.tool = "all", ...)

get_imconvert()

Arguments

lib.tool
character. Vector of libraries to look for. This can be either 'convert', 'ffmpeg', 'avconv' or a combination. Default is "all" to check for all possible libraries.

...
additional arguments. Currently not used.

Value

A character vector including all found commands or directories to the needed tools of the requested libraries. The return can serve as conv_dir input to the animate_move function.

Details

The following tools and libraries are needed by moveVis:

the convert tool of ImageMagick to support the GIF format

the ffmpeg tool of FFmpeg to support video formats

alternatively to ffmpeg, the avconv tool of libav-tools to support video formats.

It is recommmended to have both ImageMagick and one of the mentioned video libraries installed to be able to create all output formats with the animate_move() function (see argument out_format of animate_move)

If you are running Windows or macOS, use the download links below to install the required software:

ImageMagick: https://www.imagemagick.org/script/download.php)

FFmpeg from https://www.ffmpeg.org/download.html

libav from https://libav.org/download/

If you are running macOS, an installation via a package manager such as 'brew' is recommended.

If you are running Linux, execute the commands below to install the required software:

ImageMagick: sudo apt-get install imagemagick

ffmpeg: sudo apt-get install ffmpeg

libav: sudo apt-get install libav-tools

See also

Examples

#conv_dir of the animate_move() function
conv_dir <- get_libraries()