Enable multi-core usage

This function enables multi-core usage of moveVis by setting the maximum number of cores to be used.

use_multicore(n_cores = NULL, verbose = TRUE)

Arguments

n_cores
numeric, optional, number of cores to be used. If not defined, the number of cores will be detected automatically (n-1 cores will be used with n being the number of cores available).

verbose
logical, if TRUE, messages and progress information are displayed on the console (default).

Value

None. This function is used for its side effects.

Details

For most tasks, moveVis is able to use multiple cores to increase computational times through parallelization. By default, multi-core usage is disabled. This function saves the number of cores that moveVis should use to the global option "moveVis.n_cores" that can be printed using getOption("moveVis.n_cores").

moveVis uses the parallel package for parallelization.

Examples

# enable multi-core usage automatically
use_multicore()

# define number of cores manually
use_multicore(n_cores = 2)