Overview of moveVis tools

moveVis provides tools to visualize movement data of any kind, e. g by creating path animations from GPS point data. The package is under ongoing development, new functionalities are added constantly. The moveVis package is closely connected to the move package and mainly builds up on ggplot2.

Details

At the moment, the package includes the following functions:

animate_move, which can create spatial movement data animations as GIF file. Among other funtionalities, the function is able to

visualize move class point data as (multiple) movement paths,

display static basemap layers downloaded from Google Maps,

display static basemap layers provided by the user,

display dynamic, time-referenced raster data, e. g. to visualize land cover changes etc.,

compute temporal interpolations from time-referneced raster data,

create statistic plots, displaying the interaction of the individual movement paths with environmental data. ...

animate_stats, which can create animated statistic plots from movement and basemap data as GIF file.

animate_raster, which can create animated spatial plots of basemap data as GIF file.

get_libraries, a helper function to locate/download/install the extern libraries ImageMagick, FFmpeg and libav and their tools, which are needed for different output file format support.

get_formats, a helper function which returns all available output file formats (system-dependent).

See also