Tools to visualize movement data in R

moveVis provides tools to visualize movement data (e.g. from GPS tracking) and temporal changes of environmental data (e.g. from remote sensing) by creating video animations. The moveVis package is closely connected to the move package and builds up on ggplot2 grammar of graphics.

Details

The package includes the following functions, sorted by the order they would be applied to create an animation from movement data:

df2move converts a data.frame into a move or moveStack object. This is useful if you do not usually work with the move classes and your tracks are present as data.frames.

align_move aligns single and multi-individual movement data to a uniform time scale with a uniform temporal resolution needed for creating an animation from it. Use this function to prepare your movement data for animation depending on the temporal resolution that suits your data.

subset_move subsets a move or moveStack by a given time span. This is useful if you want to create a movement animation of only a temporal subset of your data, e.g. a particular day.

get_maptypes returns a character vector of available map types that can be used with frames_spatial. moveVis supports OpenStreetMaps and Mapbox basemap imergay. Alternatively, you can provide custom imagery to frames_spatial.

frames_spatial creates a list of ggplot2 maps displaying movement. Each object represents a single frame. Each frame can be viewed or modified individually. The returned list of frames can be animated using animate_frames.

frames_graph creates a list of ggplot2 graphs displaying movement-environment interaction. Each object represents a single frame. Each frame can be viewed or modified individually. The returned list of frames can be animated using animate_frames.

add_gg adds ggplot2 functions (e.g. to add layers such as points, polygons, lines, or to change scales etc.) to the animation frames created with frames_spatial or frames_graph. Instead of creating your own ggplot2 functions, you can use one of the other moveVis add_ functions:

add_labels adds character labels such as title or axis labels to animation frames created with frames_spatial or frames_graph.

add_scalebar adds a scalebar to the animation frames created with frames_spatial or frames_graph.

add_northarrow adds a north arrow to the animation frames created with frames_spatial or frames_graph.

add_progress adds a progress bar to animation frames created with frames_spatial or frames_graph.

add_timestamps adds timestamps to animation frames created with frames_spatial or frames_graph.

add_text adds static or dynamically changing text to the animation frames created with frames_spatial or frames_graph.

add_colourscale adjusts the colour scales of the animation frames created with frames_spatial and custom map imagery.

join_frames side-by-side joins the ggplot2 objects of two or more frames lists of equal lengths into a single list of ggplot2 objects per frame using plot_grid. This is useful if you want to side-by-side combine spatial frames returned by frames_spatial with graph frames returned by frames_graph.

get_frametimes extracts the timestamps associated with each frame of a list of frames created using frames_spatial or frames_graph and returns them as a vector.

suggest_formats returns a selection of suggested file formats that can be used with out_file of animate_frames on your system.

animate_frames creates an animation from a list of frames computed with frames_spatial or frames_graph.

view_spatial displays movement tracks on an interactive mapview or leaflet map.

use_multicore enables multi-core usage for computational expensive processing steps.

use_disk enables the usage of disk space for creating frames, which can prevent memory overload when creating frames for very large animations.

The majority of this functions can be used with the forward pipe operater %>%, which is re-exported by moveVis.

See also