Suggest known file formats

This function returns a selection of suggested file formats that can be used with out_file of animate_frames on your system.

suggest_formats(suggested = c("gif", "mov", "mp4", "flv", "avi", "mpeg",
  "3gp", "ogg"))

Arguments

suggested
character, a vector of suggested file formats which are checked to be known by the available renderers on the running system. By default, these are c("gif", "mov", "mp4", "flv", "avi", "mpeg", "3gp", "ogg").

Value

A subset of suggested, containing only those file formats which are known by the renderers on the running system.

See also

Examples

# find out which formats are available
suggest_formats()

# check for a particular format not listed in "suggested" that you want to use, e.g. m4v
suggest_formats("m4v")
# if "m4v" is returned, you can use this format with animate_frames